% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treePlot.R
\name{addBranch}
\alias{addBranch}
\title{color a branch}
\usage{
addBranch(tree, branch, col.branch, col.other, addTo = NULL, ...)
}
\arguments{
\item{tree}{a phylo object}

\item{branch}{a vector of node numbers labels to specify the branches to be colored. Each branch is represented by its branch node. A leaf node reprents the edge connecting the leaf and its parent.}

\item{col.branch}{a vector of colors. Its length should be one or equals to the length of \strong{branch}. If \strong{col.branch} has the same length as \strong{branch}, the branches are colored correspondingly with the \strong{col.branch}. The default is blue.}

\item{col.other}{a color for the branches other than those specified in \strong{branch}}

\item{addTo}{NULL or a plot of a phylo object.}

\item{...}{see also \code{\link[ggtree]{ggtree}}}
}
\description{
\code{addBranch} is to color a branch or some edges.
}
