% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddTreeNodeLab.R
\name{addNodeLab}
\alias{addNodeLab}
\title{add node label to the tree}
\usage{
addNodeLab(treeO, nodeLab)
}
\arguments{
\item{treeO}{an object of phylo class}

\item{nodeLab}{the label to be used for the tree nodes}
}
\value{
an object of phylo class
}
\description{
This function add labels to all tree nodes.
}
\examples{
{

library(ape)
n <- 20
Lab <- paste("Node",1:(n-1),sep="")
tipLab <- paste("T",1:n,sep="")

# entire tree
Tree <- rtree(n,tip.label= tipLab)

 # PLOT tree
ggtree(Tree,branch.length = "none")+
geom_tiplab()+
geom_text2(aes(subset=!isTip,label=label),hjust=-0.3)

new.Tree <- addNodeLab(treeO = Tree, nodeLab = Lab)

# node labels are added
ggtree(new.Tree,branch.length = "none")+
geom_tiplab()+
geom_text2(aes(subset=!isTip,label=label),hjust=-0.3)

}
}
