% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addRank.R
\name{addRank}
\alias{addRank}
\title{create rank variable}
\usage{
addRank(ModOutp, varEst, varSE, varRank, pkg = "edgeR")
}
\arguments{
\item{ModOutp}{a dataframe output from differenetial analysis package (such as: edgeR)}

\item{varEst}{the name of the column which contains the main effect estimate}

\item{varSE}{the name of the column which contains the standard error of main effect estimate}

\item{varRank}{the name to be used for the column of rank variable}

\item{pkg}{"edgeR" or "other"; if "edgeR", standard error of main effect estimates are calculated
within the function}
}
\description{
\code{addRank} is to create a rank variable which could be used later in
the tree aggregateion.
}
