% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bigTree}
\alias{bigTree}
\title{A simulated phylogenetic tree with 800 tips and 799 internal nodes}
\format{A phylo object with 800 tips and 799 internal nodes:
\describe{
  \item{Tip labels}{t1, t2, ..., t800.}
  \item{Node labels}{NULL}
}}
\usage{
bigTree
}
\description{
A random phylo object created using the function \code{\link[ape]{rtree}}
}
\keyword{datasets}
