% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tipComp.R
\name{calcDist}
\alias{calcDist}
\title{Distance calculation}
\usage{
calcDist(PointX, PointY, LineX, LineY, LineC)
}
\arguments{
\item{PointX}{the x value of a point (or points)}

\item{LineX}{the parameter of x for line (eg. if line is ax+by+c=0, then LineX = a) )}

\item{LineY}{the parameter of y for line (eg. if line is ax+by+c=0, then LineY = b)}

\item{LineC}{the parameter of y for line (eg. if line is ax+by+c=0, then LineC = c)}

\item{PointX}{the y value of a point (or points)}
}
\description{
Calculate the shortest distance between a point and a line. The distance equal to
0 if a point is on the line. Points on different sides of the line would have
different signs.
}
