% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_rate.R
\name{calcRate}
\alias{calcRate}
\title{Calculate TPR, FPR, FDR on tips level}
\usage{
calcRate(InfMat, TrueLab, type, level, mean)
}
\arguments{
\item{InfMat}{a matrix (rows: tree tips; columns: simulations.)
The entry value only includes 1 (tested to be positive) or 0 (tested to be negative).}

\item{TrueLab}{a numerical vector. Tips as names; value only include 1 (real positive)
or 0 (real negative)}

\item{type}{a charactor vector. Select from "TPR" / "FPR" / "FDR" to show which value would be
calculated."TPR" (true positive rate), "FPR" (false positive rate), "FDR" (false discovery rate)}

\item{level}{a numeric value as the threshold value of adjusted p-value. Results with value
above this level are categorized as non-significant}
}
\value{
a matrix with columns for TPR, FPR, or FDR (depends on the type selected.) and rows
for simulations
}
\description{
Calculate TPR (true positive rate), FPR (false positive rate), FDR (false discovery rate)
}
