% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classValid.R
\name{checkLSE}
\alias{checkLSE}
\title{valid leafSummarizedExperiment class}
\usage{
checkLSE(object)
}
\arguments{
\item{object}{A leafSummarizedExperiment object}
}
\value{
TRUE or a character string.
}
\description{
\code{validLSE} is to valid treeSummarizedExperiment object.
}
\keyword{internal}
