% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clusters.R
\name{ClusterForm}
\alias{ClusterForm}
\title{cut tree into clusters}
\usage{
ClusterForm(wtree, nClus)
}
\arguments{
\item{wtree}{a phylo tree to be cut;}

\item{nClus}{number of clusters}
}
\value{
a list with each element representing a cluster.
In each element, all tips in that cluster are saved.
}
\description{
\code{ClusterForm} cut tree into a defined number clusters,
and save tips in each cluster as a list
}
\examples{
\\notrun{
library(ape)
n <- 20
nodeLab <- paste("Node",1:(n-1),sep="")
tipLab <- paste("T",1:n,sep="")
# generate an ultrametric tree
Tree <- rcoal(n,tip.label= tipLab)
Tree$node.label <- nodeLab

ClusterForm(Tree,3)
}

}
