% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clusters.R
\name{ClusterInf}
\alias{ClusterInf}
\title{Information of clusters
\code{ClusterInf} summarize the total number of tips in the cluster
and the proportion of counts for each cluster calculated from the real data}
\usage{
ClusterInf(RealDat, wtree, nClus, stree)
}
\arguments{
\item{RealDat}{count table (include at least all tips of the tree)}

\item{nClus}{number of clusters}

\item{tree}{a phylo tree to be cut;}
}
\value{
a data frame

Cluster column indicates the cluster name

Freq column indicates the total number of tips in the cluster

Proportion indicates the proportions of counts for the clusters which are estimated
from the real data use \code{dirmult} from package dirmult
}
\description{
Information of clusters
\code{ClusterInf} summarize the total number of tips in the cluster
and the proportion of counts for each cluster calculated from the real data
}
\examples{
{
library(GUniFrac) # real throat data
# real data
data("throat.otu.tab")
data("throat.tree")


Lab <- paste("Node", 1:throat.tree$Nnode, sep="")
throatTree <- addNodeLab(treeO = throat.tree, nodeLab = Lab)
# prune tree at each node
smallTree <- pruneTree(wtree = throatTree)

# cluster the tree and decide on which cluster the signal is located
clusList <- ClusterForm(wtree = throatTree, nClus = 60)  # name order 1:nClus
ClusInf <- ClusterInf(RealDat = throat.otu.tab, wtree = throatTree,
nClus = 60, stree = smallTree)

}
}
