% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treePlot.R
\name{colScale}
\alias{colScale}
\title{Customize the color}
\usage{
colScale(branch, point, col.branch, col.other, col.point, legend.label,
  legend.title, size.line.legend, legend)
}
\arguments{
\item{branch}{A vector of node numbers labels to specify the branches to be colored. Each branch is represented by its branch node. A leaf node reprents the edge connecting the leaf and its parent.}

\item{point}{A vector of node numbers or node labels to specify the locations to add points in the tree.}

\item{col.branch}{A vector of colors. Its length should be one or equals to the length of \strong{branch}. If \strong{col.branch} has the same length as \strong{branch}, the branches are colored correspondingly with the \strong{col.branch}. The default is blue.}

\item{col.other}{A color for the branches other than those specified in \strong{branch}}

\item{col.point}{A color for the \strong{point}. It has length equal to one.}

\item{legend.label}{A list with three members: "col.branch", "col.other", and "col.point". The elements order in each member matches with the corresponding argument \strong{col.branch}, \strong{col.other} and \strong{col.point}, and will display in the legend. See Examples.}

\item{legend.title}{A vector to specify the title of the legend. It must be named with "branch" and "point" to match with the argument \strong{branch} and \strong{point}.}

\item{size.line.legend}{the line size shown in the legend for \strong{branch}}

\item{legend}{TRUE or FALSE. Default is FALSE. If TRUE, the legend is created.}
}
\description{
\code{colScale} customizes the color scale.
}
