% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cytofCluster}
\alias{cytofCluster}
\title{The information about clusters of the AML-sim data}
\format{A data frame of cluster information:
\describe{
  \item{cluster}{The name of cluster}
  \item{n_cells}{The number of cells in the cluster}
  \item{n_spikein}{The number of spike-in cells in the cluster}
  \item{prop_spikein}{The proportion of spike-in cells in the cluster}
  \item{DA}{whether the cluster is differential abundant between the healthy
  sample and the CN sample}
}}
\usage{
cytofCluster
}
\description{
A data frame provides information about clusters. Each cluster corresponds to
a row in the \code{\link{cytofCount}}.
}
\keyword{datasets}
