% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cytofSample}
\alias{cytofSample}
\title{The information about samples of the AML-sim data}
\format{5 samples are collected from 5 healthy subjects. Each sample is
  equally split into two subsamples, one of which spike in CN cells to
  simulate samples from patients.
\describe{
  \item{group}{The group that a sample collected from: healthy or CN.}
  \item{patient}{The patient that the sample collected from. }
}}
\usage{
cytofSample
}
\description{
A data frame provides information about samples. Each sample corresponds to
a column in the \code{\link{cytofCount}}.
}
\keyword{datasets}
