% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cytofTree}
\alias{cytofTree}
\title{The hierarchical tree of the cells in the AML-sim data}
\format{A hclust object with 900 leaves. Each leaf is a cell cluster.}
\usage{
cytofTree
}
\description{
A hclust object which records the hierarchical structure of the cells in the
AML-sim data.
}
\keyword{datasets}
