% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cytofTruth}
\alias{cytofTruth}
\title{The true signal in the AML-sim data}
\format{A data frame of two columns:
\describe{
  \item{cluster}{The cluster label}
   \item{true}{Whether the cluster is truly differential abundant}
}}
\usage{
cytofTruth
}
\description{
This data provides the differential abundance information of clusters in the
AML-sim data
}
\keyword{datasets}
