% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doTable.R
\name{doCount}
\alias{doCount}
\title{generate a count table}
\usage{
doCount(FC, data, nSam, mu, size, n)
}
\arguments{
\item{FC}{a numeric vector. Fold change}

\item{data}{a matrix or data frame. the count table from real data}

\item{nSam}{a vector of length two. The number of samples in two different groups or conditions.}

\item{mu, size}{the parameters of negative binomial distribution (see mu and size in \code{\link[stats]{rnbinom}}.)}

\item{n}{a numeric value. the number of count table generated. If above one, the output is a list of matrice.}
}
\value{
a matrix or a list of matrice
}
\description{
\code{doCount} is to generate a count table given some available information. Here, the information includes the fold change between two conditions, the parameters of negative binomial distritubion (which the sample library size follows), a data table from real data (to estimate the proportion of each entity in the sample), and the number of samples in two different groups or conditions.
}
