% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewSim.R
\name{.findParallel}
\alias{.findParallel}
\title{find parallel nodes}
\usage{
.findParallel(tree, input, use.alias = FALSE)
}
\arguments{
\item{tree}{A phylo object.}

\item{input}{A numeric or character vector. Node labels or node numbers.}

\item{use.alias}{A logical value, TRUE or FALSE. The default is FALSE, and
the node label would be used to name the output; otherwise, the alias of
node label would be used to name the output. The alias of node label is
created by adding a prefix \code{"Node_"} to the node number if the node is
an internal node or adding a prefix \code{"Leaf_"} if the node is a leaf
node.}
}
\value{
A vector of nodes. The numeric value is the node number, and the
  vector name is the corresponding node label. If a node has no label, it
  would have NA as name when \code{use.alias = FALSE}, and have the alias of
  node label as name when \code{use.alias = TRUE}.
}
\description{
\code{.findParallel} is to find target nodes that could build up a tree with
some specified nodes. The returned target nodes is the combination that has
the minimum number of nodes. In other words, a tree is cut in a way that the
number of branches is the minimum in all possibilities and the specified
branches are obtained. A branch is represented by its branch node. A leaf
node represents the edge connecting the leaf and its parent.
}
\examples{
# data(tinyTree)
# .findParallel(tree = tinyTree, input = 12,
#             use.alias = FALSE)


}
\keyword{internal}
