% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_offspring.R
\name{elementFind}
\alias{elementFind}
\title{Find offsprings}
\usage{
elementFind(ancestor, stree, only.Tip = TRUE)
}
\arguments{
\item{ancestor}{A charactor vector includes the label of nodes and tips.
If tips are included, themselves are returned as offsprings}

\item{stree}{a list of all the subtrees of a full phylogenetic tree}

\item{only.Tip}{if TRUE, only descendant tips are returned; otherwise,
both tips and nodes are returned}
}
\value{
a character vector which includes labels of offspring tips
}
\description{
Find offsprings for nodes.
}
\examples{
\dontrun{
library(ape)
library(ggtree)

n <- 20
Lab <- paste("Node",1:(n-1),sep="")
tipLab <- paste("T",1:n,sep="")

# entire tree
set.seed(1)
Tree <- rtree(n,tip.label= tipLab)
new.Tree <- addNodeLab(treeO = Tree, nodeLab = Lab)
sTree <- pruneTree(new.Tree)

# PLOT tree
ggtree(new.Tree,branch.length = "none")+
geom_tiplab()+
geom_text2(aes(subset=!isTip,label=label),hjust=-0.3)

# find offspring
offsp1 <- FindOffspring(ancestor = "Node14", stree = sTree,
only.Tip = TRUE)
offsp2 <- FindOffspring(ancestor = "Node14", stree = sTree,
only.Tip = FALSE)
}
}
