% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdr.R
\name{fdr}
\alias{fdr}
\title{calculate false discovery rate (fdr) on a tree structure}
\usage{
fdr(tree, truth, found, level = c("node", "leaf"), direction = FALSE)
}
\arguments{
\item{tree}{a phylo object}

\item{truth}{Nodes have signals (eg. differentally abundant at different experimental conditions.). If the signals are in different directions (up and down), then provide the nodes as a list of two members (one up and one down). One could provide either node number or node label. \strong{Note:} When fdr at node level is required and only leaf nodes with signal are provided, then the internal nodes which are shared and only shared by the provied leaf nodes (signal in the same direction) will be found out and used with the leaf nodes in the fdr calculation; when fdr at leaf level is required and the given nodes have internal nodes, then the descendant leaf nodes will be found out and used in the fdr calculation.}

\item{found}{Nodes have been found to have signal (eg. differentally abundant at different experimental conditions). If the signals are in different directions (up and down), then provide the nodes as a list of two members (one up and one down). One could provide either node number or node label. \strong{Note:} When fdr at node level is required, then the descendant nodes of the provied nodes (include themselves) will be found out and used in the fdr calculation; when fdr at leaf level is required, then the descendant leaf nodes will be found out and used in the fdr calculation.}

\item{level}{if "leaf", false discovery rate is calculated at leaf level; if "node", it is calculated at node level.}

\item{direction}{TRUE or FALSE. Default is FALSE. If TRUE, the signal direction is taken into account; the argument \strong{truth} and \strong{found} should both be a list of two members and the order of directions should match.}
}
\value{
a false discovery rate
}
\description{
\code{fdr} is to calculate the false discovery rate (fdr) on a tree structure at leaf or node level.
}
\examples{

library(ggtree)
data("tinyTree")
ggtree(tinyTree) + geom_text2(aes(label = node))

# Assume there are two branches with signal: one with root node 16 (signal direction: up) and the other with root node 13 (signal direction: down)

# ------------ ignore direction -------------------
 # two branches are found to have signal: node 17 and node 14
 # fdr at the tip level
fdr1 <- fdr(tree = tinyTree, truth = c(16, 13),
            found = c(17, 14), level = "leaf")
 # fdr at the node level
fdr2 <- fdr(tree = tinyTree, truth = c(16, 13),
            found = c(17, 14), level = "node")


# ------------ direction matters-------------------
# two branches have signal: node 16 (up) and node 13 (down)
# two branches are found to have signal: node 17 (up) and node 14 (down)

 # make sure the direction order in the list for truth and found is the          same
(fdr3 <- fdr(tree = tinyTree, truth = list(16, 13),
            found = list(16, 13), level = "leaf",
            direction = TRUE))  # correct order

(fdr4 <- fdr(tree = tinyTree, truth = list(16, 13),
            found = list(13, 16), level = "node",
            direction = TRUE)) # wrong order



}
