% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindCAND.R
\name{FindCAND}
\alias{FindCAND}
\title{list all possible best results
\code{FindCAND} is to find results which are close to the truth}
\usage{
FindCAND(iNode, stree, trueN)
}
\arguments{
\item{iNode}{a vector of inferred nodes}

\item{stree}{a list of phylo object}

\item{trueN}{a vector of true nodes (tips)}
}
\value{
a list which lists all found results. Each element includes a vector
of kept inferred nodes (tips)
}
\description{
list all possible best results
\code{FindCAND} is to find results which are close to the truth
}
\examples{
{
library(ape)
n <- 20
Lab <- paste("Node",1:(n-1),sep="")
tipLab <- paste("T",1:n,sep="")
# entire tree
set.seed(1)
Tree <- rtree(n,tip.label= tipLab)
Tree <- addNodeLab(treeO = Tree, nodeLab = Lab)
# Subtrees
sTree <- pruneTree(Tree)
# PLOT
ggtree(Tree,branch.length = "none") +
  geom_text2(aes(subset=!isTip, label=label), hjust=-.3) +
  geom_tiplab()

# inferred nodes & tips
infNodes <- c("Node2", "Node3", "Node4", "Node7",
              "Node8", "Node9", "Node10", "Node11", "Node12",
              "Node18", "Node15", "T12")
trueNode <- c("Node9","Node17")
# final result
rn <- FindCAND(iNode = infNodes, stree = sTree, trueN)
}

}
