% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_offspring.R
\name{FindOffspring}
\alias{FindOffspring}
\title{Find offsprings}
\usage{
FindOffspring(ancestor, stree)
}
\arguments{
\item{ancestor}{A charactor vector includes the label of nodes and tips.
If tips are included, themselves are returned as offsprings}

\item{stree}{a list of all the subtrees of a full phylogenetic tree}
}
\value{
a character vector which includes labels of offspring tips
}
\description{
Find offsprings for nodes.
}
\examples{
\dontrun{
rnorm(10)
}
}
