% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tipComp.R
\name{FreqMat}
\alias{FreqMat}
\title{Calculate the frequence of rows with same entries in a matrix}
\usage{
FreqMat(Mat)
}
\arguments{
\item{Mat}{a matrix}
}
\value{
a vector includes the frequency value of each row.
}
\description{
FreqMat calculate the frequency of rows with same values in a matrix
}
