% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupMat.R
\name{groupMat}
\alias{groupMat}
\title{convert inferred results into clusters}
\usage{
groupMat(vectNT, stree)
}
\arguments{
\item{vectNT}{a vector includes inferred nodes (represent branch) and inferred tips}

\item{stree}{a list of phylo object}
}
\value{
a vector which assigned a cluster number to each tip of the tree. In total, there are
n+1 clusters if the inferred result has length n.
}
\description{
\code{groupMat} converts the inferred nodes into clusters by combining the tree information.
Each inferred node represent a branch. The root node of the branch is the inferred node. All
tips in a branch are grouped as a cluster. If the inferred result also include tips, then each
inferred tip is a cluster. The tips not included in the inferred branch or inferred tips are
grouped as one cluster. Therefore, if the inferred result includes n elements (include both tips
and branches), then the final cluster number is n+1.
}
\examples{
{
library(ape)
n <- 20
Lab <- paste("Node",1:(n-1),sep="")
tipLab <- paste("T",1:n,sep="")
# entire tree
set.seed(1)
Tree <- rtree(n,tip.label= tipLab)
Tree <- addNodeLab(treeO = Tree, nodeLab = Lab)
# Subtrees
sTree <- pruneTree(Tree)
# PLOT
ggtree(Tree,branch.length = "none") +
  geom_text2(aes(subset=!isTip, label=label), hjust=-.3) +
  geom_tiplab()

# inferred nodes & tips
infNodes <- c("Node2", "T12")

# final result
cc <- groupMat(vectNT = infNodes, stree = sTree)

}
}
