% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doTable.R
\name{infLoc}
\alias{infLoc}
\title{provide the information of two branches}
\usage{
infLoc(tree, data, from.A, from.B)
}
\arguments{
\item{tree}{a phylo object}

\item{data}{count table (a matrix or a data frame). It is a real data. It has tree leaves in rows and samples from different conditions in  columns.}

\item{from.A, from.B}{the branch node labels of Branch A, B.}
}
\value{
a data frame of one row
}
\description{
\code{infLoc} is to give information of two branches about the count proportion and the number of leaves
}
