% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClass.R
\docType{class}
\name{internal_rowData-class}
\alias{internal_rowData-class}
\title{int_rowData: A virtual class
-----------------------------------------------------------------------------
A virtual class to be assigned to a column of \code{rowData} in a
\code{treeSummarizedExperiment} object. The columns with this class could be
optionally exported or hidden when using \code{rowData}.}
\description{
int_rowData: A virtual class
-----------------------------------------------------------------------------
A virtual class to be assigned to a column of \code{rowData} in a
\code{treeSummarizedExperiment} object. The columns with this class could be
optionally exported or hidden when using \code{rowData}.
}
\keyword{internal}
