% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewSim.R
\name{isTip}
\alias{isTip}
\title{To test whether specified nodes are leaf nodes}
\usage{
isTip(tree, input)
}
\arguments{
\item{tree}{A phylo object.}

\item{input}{A numeric or character vector. Node labels or node numbers.}
}
\value{
a logical vector with the same length as input
}
\description{
\code{isTip} is to test wheter some specified nodes are leaf nodes of a tree.
}
\keyword{internal}
