% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/classAccessor.R
\docType{methods}
\name{assayNames,leafSummarizedExperiment-method}
\alias{assayNames,leafSummarizedExperiment-method}
\alias{assayNames<-,leafSummarizedExperiment,ANY-method}
\alias{assays,leafSummarizedExperiment-method}
\alias{assays<-,leafSummarizedExperiment,ANY-method}
\alias{rowData,leafSummarizedExperiment-method}
\alias{rowData<-,leafSummarizedExperiment-method}
\alias{colData,leafSummarizedExperiment-method}
\alias{colData<-,leafSummarizedExperiment,ANY-method}
\alias{metadata,leafSummarizedExperiment-method}
\alias{leafSummarizedExperiment-accessor}
\title{Accessor functions for leafSummarizedExperiment}
\usage{
\S4method{assayNames}{leafSummarizedExperiment}(x, ...)

\S4method{assayNames}{leafSummarizedExperiment,ANY}(x, ...) <- value

\S4method{assays}{leafSummarizedExperiment}(x, ..., withDimnames = TRUE)

\S4method{assays}{leafSummarizedExperiment,ANY}(x, ...,
  withDimnames = TRUE) <- value

\S4method{rowData}{leafSummarizedExperiment}(x, ...)

\S4method{rowData}{leafSummarizedExperiment}(x, ...) <- value

\S4method{colData}{leafSummarizedExperiment}(x, ...)

\S4method{colData}{leafSummarizedExperiment,ANY}(x, ...) <- value

\S4method{metadata}{leafSummarizedExperiment}(x)
}
\arguments{
\item{x}{A leafSummarizedExperiment object}

\item{...}{For assay, ... may contain withDimnames, which is forwarded to
assays. For rowData, arguments passed through ... are forwarded to mcols.
For cbind, rbind, ... contains leafSummarizedExperiment objects to be
combined. For other accessors, ignored.}

\item{value}{An object of a class specified in the S4 method signature. See
\code{\link[SummarizedExperiment]{SummarizedExperiment}} for more details.}

\item{withDimnames}{A logical(1), indicating whether dimnames should be
applied to extracted assay elements. Setting withDimnames=FALSE increases
the speed and memory efficiency with which assays are extracted.
withDimnames=TRUE in the getter assays<- allows efficient complex
assignments (e.g., updating names of assays, names(assays(x,
withDimnames=FALSE)) = ... is more efficient than names(assays(x)) = ...);
it does not influence actual assignment of dimnames to assays.}
}
\value{
Elements from \code{leafSummarizedExperiment}.
}
\description{
Accessor functions to extract different elements from
\strong{leafSummarizedExperiment} object.
}
\seealso{
\code{\link{treeSummarizedExperiment}}
  \code{\link{treeSummarizedExperiment-accessor}}
  \code{\link{leafSummarizedExperiment}}
  \code{\link[SummarizedExperiment]{SummarizedExperiment-class}}
}
\author{
Ruizhu HUANG
}
