% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClass.R
\docType{class}
\name{leafSummarizedExperiment-class}
\alias{leafSummarizedExperiment-class}
\alias{.lse}
\title{An S4 class leafSummarizedExperiment}
\description{
The class \strong{leafSummarizedExperiment} is an extension class of standard
\code{\link[SummarizedExperiment]{SummarizedExperiment}} class. It is
designed to store rectangular data like a \code{matrix} for entities (e.g.,
microbes or cell types), and the hiearchical structure information of
entities. The annotations on the rows and columns of the rectangular data are
strored in \code{rowData} and \code{colData}, respectively. Each row of the
rectangular data could be mapped to a leaf node of the tree via the rownames
of \strong{rowData} or a column named as \strong{nodeLab} in
\strong{rowData}. For example, we could include abundance count of microbes
collected from different samples in the \strong{assays} and the phylogenetic
tree of microbes in the \strong{metadata}. Each row of matrix-like data in
\strong{assays} corresponds to one microbial specie, whose information is
given in \strong{rowData}. We could find the label of its corresponding leaf
node in column \strong{nodeLab}, or in the rownames of \strong{rowData}. Each
column of matrix-like data is a sample, whose information is stored in
\strong{colData}. The class \strong{leafSummarizedExperiment} has four slots
\strong{assays}, \strong{rowData} \strong{colData} and \strong{metadata} as
\code{\link[SummarizedExperiment]{SummarizedExperiment}} class. More details
about these four slots could be found in
\code{\link[SummarizedExperiment]{SummarizedExperiment-class}}.
}
\details{
\strong{leafSummarizedExperiment} has more restrictions in data structure
than \strong{SummarizedExperiment}.
In \strong{leafSummarizedExperiment} class, we have restrictions as below.
\itemize{
\item It's not allowed to have empty \strong{assays}.

\item A phylo object named as \strong{tree} should be contained in
\strong{metadata}. The labels of tree leaf nodes are unique.

\item Each rows of matrix-like data in \strong{assays} could be mapped to a
leaf node of the tree. The label of leaf node is provided in a column
\strong{nodeLab} in \strong{rowData}, or provides as the rownames of
\strong{rowData}. If both provided, the information in column \strong{nodeLab} is
used. }
}
\section{Constructor}{

See \code{\link{leafSummarizedExperiment-constructor}} for constructor
functions.
}

\section{Accessor}{

See \code{\link{leafSummarizedExperiment-accessor}} for accessor functions.
}

