% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClass.R
\name{leafSummarizedExperiment-constructor}
\alias{leafSummarizedExperiment-constructor}
\alias{leafSummarizedExperiment}
\title{Construct a leafSummarizedExperiment object}
\usage{
leafSummarizedExperiment(tree, ...)
}
\arguments{
\item{tree}{A phylo object.}

\item{...}{For \code{SummarizedExperiment}, S4 methods \code{list}
    and \code{matrix}, arguments identical to those of the
    \code{SimpleList} method.

    For \code{rowRanges}, ignored.

  }
}
\value{
A leafSummarizedExperiment object
}
\description{
\code{leafSummarizedExperiment} is the constructor of the
\strong{leafSummarizedExperiment} class.
}
\examples{
# tree
data("tinyTree")

# the count table
count <- matrix(rpois(100, 50), nrow = 10)
rownames(count) <- c(tinyTree$tip.label)
colnames(count) <- paste("C_", 1:10, sep = "_")

# the sample information
sampC <- data.frame(condition = rep(c("control", "trt"),
                                    each = 5),
                    gender = sample(x = 1:2, size = 10,
                                    replace = TRUE))
rownames(sampC) <- colnames(count)

# build a leafSummarizedExperiment object
lse <- leafSummarizedExperiment(tree = tinyTree,
                                assays = list(count),
                                colData = sampC)

}
\seealso{
\code{\link{leafSummarizedExperiment-class}}
  \code{\link{treeSummarizedExperiment-class}}
  \code{\link[SummarizedExperiment]{SummarizedExperiment-class}}
}
\author{
Ruizhu HUANG
}
