% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchReal.R
\name{matchReal}
\alias{matchReal}
\title{match inferred results with the truth}
\usage{
matchReal(infN, trueN, stree, decreasing = FALSE)
}
\arguments{
\item{infN}{a vector of inferred results (tips & nodes)}

\item{trueN}{a vector of true results (tips & nodes)}

\item{stree}{a list of phylo object}

\item{decreasing}{logical value; indicate the sort order}
}
\description{
\code{matchReal} assigns the inferred nodes (or tips) to the corresponding true
signal nodes. For each true signal nodes, the assigned inferred nodes are sorted
in increasing order according to the number of their descendant tips
}
\examples{
{
library(ape)
library(ggtree)
n <- 20
Lab <- paste("Node",1:(n-1),sep="")
tipLab <- paste("T",1:n,sep="")
# entire tree
set.seed(1)
Tree <- rtree(n,tip.label= tipLab)
new.Tree <- addNodeLab(treeO = Tree, nodeLab = Lab)
sTree <- pruneTree(new.Tree)
# PLOT tree
ggtree(new.Tree,branch.length = "none")+
  geom_tiplab()+
  geom_text2(aes(subset=!isTip,label=label),hjust=-0.3)

# true nodes
tN <- c("Node13","Node2")

# inferred nodes & tips
iN <- c("Node14","Node16","Node2")

mList <- matchReal(infN = iN, trueN = tN, stree = sTree)
}

}
