% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClass.R
\docType{class}
\name{matrixDataframe-class}
\alias{matrixDataframe-class}
\title{A virtual class to combine "matrix", "data.frame".}
\description{
A virtual class to combine "matrix", "data.frame".
}
\keyword{internal}
