% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simuCount.R
\name{nodeCount}
\alias{nodeCount}
\title{Aggregate the count of tips to the nodes}
\usage{
nodeCount(tipTable, wtree, stree)
}
\arguments{
\item{tipTable}{the tip count table}

\item{wtree}{a tree (phylo class)}

\item{stree}{a list of phylo class;
a list of subtrees cut at each node of wtree}
}
\value{
a count table (matrix class) with row representing a tip /node and
columns representing samples. Samples are grouped into two conditions.
Samples in condition 1 and condition 2 are named corresponding with
the start string "C1_" and string "C2_"
}
\description{
This function is to calculate the counts at each node. The node count is
the sum of the counts of its descendant tips.
}
