% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodeValue.R
\name{nodeValue.A}
\alias{nodeValue.A}
\title{Calculate entity values for internal nodes}
\usage{
nodeValue.A(tree, data, fun = sum)
}
\arguments{
\item{tree}{A phylo object}

\item{data}{A matrix or data frame.}

\item{fun}{A function.}
}
\value{
A count table (matrix class) with a row representing a node and a
  column representing a sample.
}
\description{
\code{nodeValue} calculates the entity value at each internal node.
}
\author{
Ruizhu Huang
}
\keyword{internal}
