% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodeValue.R
\name{nodeValue.B}
\alias{nodeValue.B}
\title{Calculate entity value for each internal node}
\usage{
nodeValue.B(tree, data, fun = sum)
}
\arguments{
\item{tree}{A phylo object}

\item{data}{A treeSummarizedExperiment}

\item{fun}{A function}
}
\value{
A treeSummarizedExperiment.
}
\description{
\code{nodeValue.B} calculates value, such as, count, for each internal node.
}
\author{
Ruizhu Huang
}
\keyword{internal}
