% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parEstimate.R
\name{parEstimate}
\alias{parEstimate}
\title{Parameter estimation in Dirichlet-multinomial distribution}
\usage{
parEstimate(data)
}
\arguments{
\item{data}{A matrix or leafSummarizedExperiment. Samples in the columns and
entities in the rows.}
}
\value{
A list including \dQuote{pi} and \dQuote{theta}
}
\description{
\code{parEstimate} is a wrapper function generated from the function
\code{\link[dirmult]{dirmult}} with default setting on \code{init},
\code{initscalar}, \code{epsilon}, \code{trace} and \code{mode}. It allows
the input \code{data} to accept \code{matrix} or
\code{leafSummarizedExperiment} and output the value \code{pi} and
\code{theta}.
}
\examples{
\dontrun{
library(GUniFrac)
data("throat.otu.tab")
throat.par <- parEstimate(data = t(throat.otu.tab))
}

}
\author{
Ruizhu Huang
}
