% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathTree.R
\name{pathTree}
\alias{pathTree}
\title{Convert a tree into multiple paths}
\usage{
pathTree(tree)
}
\arguments{
\item{tree}{a phylo object}
}
\description{
\code{treePath} convert a phylo object tree into multiple paths. Each starts with a tip, connects all internal nodes, and end with the root.
}
\examples{

data(tinyTree)
library(ggtree)
ggtree(tinyTree) + geom_text2(aes(label = node))

(paths <- pathTree(tinyTree))

}
