% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pruneTree.R
\name{pruneTree}
\alias{pruneTree}
\title{prune tree at each node}
\usage{
pruneTree(tree, message = FALSE)
}
\arguments{
\item{tree}{an object of phylo class}

\item{message}{a logical value, TRUE or FALSE. If default (FALSE), the process isn't printed out.}
}
\value{
a list of phylo object. If the tree ('tree') provides node and tip labels, the labels are passed to its subtrees; otherwise, the labels are generated automatically with a prefix 'Node_' for internal nodes and 'Tip_' for leaves with the corresponding numbers in the edge component of phylo object. (tree$edge)
}
\description{
\code{pruneTree} prune the tree at each node.
}
\examples{

data(tinyTree)
small <- pruneTree(tree = tinyTree, message = TRUE)
}
