% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{res_table}
\alias{res_table}
\title{AML-sim data}
\format{A data frame of cell counts collected from healthy and
  cytogenetically normal (CN) samples:
\describe{
  \item{cluster}{The cluster labels}
  \item{p_vals}{The nominal p values}
  \item{p_adj}{The adjusted p values}
  \item{n_cells}{The number of cells in the clusters}
  \item{n_spikein}{The number of spike-in cells}
  \item{prop_spikein}{The proportion of spike-in cells}
  \item{true}{Whether the cluster is truly differential abundant}
  \item{healthy_H1}{The counts from the first healthy sample}
  \item{healthy_H2}{The counts from the second healthy sample}
  \item{healthy_H3}{The counts from the third healthy sample}
  \item{healthy_H4}{The counts from the forth healthy sample}
  \item{healthy_H5}{The counts from the fifth healthy sample}
  \item{CN_H1}{The counts from the first CN sample}
  \item{CN_H2}{The counts from the second CN sample}
  \item{CN_H3}{The counts from the third CN sample}
  \item{CN_H4}{The counts from the forth CN sample}
  \item{CN_H5}{The counts from the fifth CN sample}
}}
\usage{
res_table
}
\description{
A semi-simulated count table generated by computationally "spiking in" small
percentages of AML blast cells into healthy BMMCs (created by Lukas Weber)
}
\keyword{datasets}
