% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmDesc.R
\name{rmDesc}
\alias{rmDesc}
\title{Remove nodes which are the descendants of other nodes}
\usage{
rmDesc(tree, node)
}
\arguments{
\item{tree}{A list of phylo objects.}

\item{node}{A vector of node labels or node numbers}
}
\value{
A vector of node numbers
}
\description{
\code{rmDesc} removes branches, which are the subbranches of the others.
}
\examples{
{
data(tinyTree)
library(ggtree)
# PLOT
ggtree(tinyTree,branch.length = 'none') +
  geom_text2(aes(label=node), hjust=-.3)

# find the shared nodes from the tree plot
aV <- c(5, 4, 18, 3, 2)
# final result
(rn <- rmDesc(node = aV, tree = tinyTree))
}

}
\author{
Ruizhu Huang
}
