% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmSuper.R
\name{rmSuper}
\alias{rmSuper}
\title{remove super branches}
\usage{
rmSuper(vectNT, stree)
}
\arguments{
\item{vectNT}{a vector; a branch is represented by its root node.}

\item{stree}{a list of phylo object.}
}
\description{
\code{rmSuper} is to remove the branches which have sub-branches or tips.
If many branches (represented by its root node) and tips are given as a vector,
this function could find and remove the branch which has at least one
subbranch or tip existed in the given vector.
}
\examples{
{
library(ape)
n <- 20
Lab <- paste("Node",1:(n-1),sep="")
tipLab <- paste("T",1:n,sep="")
# entire tree
set.seed(1)
Tree <- rtree(n,tip.label= tipLab)
Tree <- addNodeLab(treeO = Tree, nodeLab = Lab)
# Subtrees
sTree <- pruneTree(Tree)
# PLOT
ggtree(Tree,branch.length = "none") +
  geom_text2(aes(subset=!isTip, label=label), hjust=-.3) +
  geom_tiplab()
# find the shared nodes from the tree plot
aV <- c("Node5","Node4","Node2","T11","Node18","Node12","T3")
# final result
rn <- rmSuper(vectNT = aV, stree = sTree)
}

}
