% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runEdgeR.R
\name{runEdgeR}
\alias{runEdgeR}
\title{Differential analysis using edgeR package}
\usage{
runEdgeR(countTab, nSam, isTip, isAnalyze, prior.count, normalize = TRUE,
  method = "TMM")
}
\arguments{
\item{countTab}{a numeric matrix indicates the count table of entities (eg. genes or OTUs) in samples under different conditions. Each row represents an entity and each column represents a sample.}

\item{nSam}{a numeric vector. Each element represents the number of samples in one condition. For example, if we have n1 samples in condition 1 and n2 samples in condition 2, then the \strong{nSam} is (n1,n2).}

\item{isTip}{a logical vector. It has equal length to the number of rows in \strong{countTab}. If TRUE, the corresponding row contributes to the library size of a sample; otherwise, it doesn't. A provided count table might include rows for leaves and internal nodes. Only the counts from leaf nodes are really measured and should be included to do library size normalisation.}

\item{isAnalyze}{a logical vector with length equal to the row number of the countTab. This is to indicate the row used for differential analysis.}

\item{prior.count}{a numeric value; the average prior count added to each observation to compute estimated coefficients for a NB glm in such a way that the log-fold-changes are shrunk towards zero (see \code{\link[edgeR]{predFC}}.}

\item{normalize}{a logical value; indicate whether to do library size normalization. If TRUE, TMM( weighted trimmed mean of M-values is applied); otherwise, raw library size is used.}

\item{method}{see \strong{method} in \code{\link[edgeR]{calcNormFactors}}}
}
\value{
a data frame containing the elements
 logFC, the log-abundance ratio, i.e. fold change, for each tag in the two groups being compared logCPM, the log-average concentration/abundance for each tag in the two groups being compared
 PValue, exact p-value for differential expression using the NB model

 FDR, the p-value adjusted for multiple testing as found using p.adjust using the method BH.

 predLFC, the predictive log2 fold changes (SEE \code{\link[edgeR]{predFC}})

 estimate, the estimated coefficient logFC = estimate / log(2)

 tag.disp, the tagwise dispersion

 waldAP, the adjusted p-value from Wald test ('BH' method; Benjamini & Hochberg (1995))
}
\description{
\code{runEdgeR} is to do differential analysis for the data using edgeR package
}
