% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selNode.R
\name{selNode}
\alias{selNode}
\title{Select branches}
\usage{
selNode(tree, data, minTip = 0, maxTip = Inf, minPr = 0, maxPr = 1,
  skip = NULL, all = FALSE)
}
\arguments{
\item{tree}{A phylo object}

\item{data}{A count table from the real data or a list output from
\code{parEstimate}.}

\item{minTip}{the minimum number of leaves in the selected branch}

\item{maxTip}{The maximum number of leaves in the selected branch}

\item{minPr}{The minimum count proportion of the selected branch in a sample.
A value between 0 and 1.}

\item{maxPr}{The maximum count proportion of the selected branch in a sample.
A value between 0 and 1.}

\item{skip}{A character vector of node labels. These nodes are not the
descendants or the ancestors of the selected branch.}

\item{all}{TRUE or FALSE. Default is FALSE. If FALSE, the branch node of a
branch, which meet the requirements and has the minimum count proportion,
is returned; otherwise branch nodes of all branches meeting the
requirements are returned.}
}
\value{
The node whose descendant branch has the lowest proportion
}
\description{
\code{selNode} selects branches meeting the specified criteria in the number
of leaves and the count proportion.
}
\examples{
\dontrun{
data("cytofTree")
data("cytofCount")
set.seed(1)
library(ape)
tree <- as.phylo(cytofTree)
sN <- selNode(tree = tree, minTip = 20,
maxTip = 50, minPr = 0.02, maxPr = 0.03,
skip = NULL, data = cytofCount, all = TRUE)
}

}
\author{
Ruizhu Huang
}
