% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signalFind.R
\name{signalFind}
\alias{signalFind}
\title{find the node shared by the specified nodes}
\usage{
signalFind(node, tree, label = FALSE)
}
\arguments{
\item{node}{a vector of node numbers or node labels}

\item{tree}{a tree (phylo object)}

\item{label}{a logical value. If TRUE, the selected node label is output; otherwise the node number is output}
}
\value{
the label of the shared node
}
\description{
\code{signalFind} is to find nodes which are the ancestors shared by the specified nodes.  The descendant leaves of the found nodes are also the descendant leaves of the specified nodes.
}
\examples{

data(tinyTree)
# PLOT tree
ggtree(tinyTree,branch.length = 'none')+
geom_text2(aes(label = label))

## find the node shared by provided node labels
signalFind(node = c('t4','t9'), tree = tinyTree,
 label = TRUE)

signalFind(node = c('t10','Node_18', 't8'), tree = tinyTree,
 label = TRUE)

# -----------------------------------------------------
ggtree(tinyTree,branch.length = 'none')+
geom_text2(aes(label = node))

## find the node shared by provided node numbers
signalFind(node = c(2, 3), tree = tinyTree)
signalFind(node = c(2, 3, 116), tree = tinyTree)

}
