% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simCount}
\alias{simCount}
\title{the count table of AML-sim data}
\format{A data frame of cell counts collected from healthy and cytogenetically normal (CN) samples:
\describe{
  \item{healthy_H1}{The counts gained from the first healthy sample}
  \item{healthy_H2}{The counts gained from the second healthy sample}
  \item{healthy_H3}{The counts gained from the third healthy sample}
  \item{healthy_H4}{The counts gained from the forth healthy sample}
  \item{healthy_H5}{The counts gained from the fifth healthy sample}
  \item{CN_H1}{The counts gained from the first CN sample}
  \item{CN_H2}{The counts gained from the second CN sample}
  \item{CN_H3}{The counts gained from the third CN sample}
  \item{CN_H4}{The counts gained from the forth CN sample}
  \item{CN_H5}{The counts gained from the fifth CN sample}
}}
\usage{
simCount
}
\description{
A semi-simulated count table generated by computationally "spiking in" small percentages of AML blast cells into healthy BMMCs (created by Lukas Weber)
}
\keyword{datasets}
