% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simTree}
\alias{simTree}
\title{The hiearchical tree of the cells}
\format{A hclust object with 900 leaves. Each leaf is a cell cluster.}
\usage{
simTree
}
\description{
A hclust object which records the hiearchical structure of the cells.
}
\keyword{datasets}
