% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simTruth}
\alias{simTruth}
\title{The true signal in AML-sim data}
\format{A data frame of two columns:
\describe{
  \item{cluster}{The label of cluster}
   \item{true}{whether the cluster is truly differential abundant}
}}
\usage{
simTruth
}
\description{
This data provides the differential abundance information of clusters in AML-sim data
}
\keyword{datasets}
