% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simuCount.R
\name{simuCount}
\alias{simuCount}
\title{simulate a count table based on the signals on a tree}
\usage{
simuCount(RealDat, wtree, nClus, nSam, muNB, sizeNB, swapClus,
  diffClus = NULL, FC = NULL)
}
\arguments{
\item{RealDat}{real data used to estimate the proportion of tips on a tree}

\item{wtree}{a tree (phylo class) where signal branch could be specified}

\item{nClus}{the number of clusters which we would like to cut the tree
and finally select some clusters to add signal}

\item{nSam}{the sample number for each condition}

\item{muNB}{mean for negative binomial (mu argument in stats::rnbinom)
(to sample library size)}

\item{sizeNB}{target for number of successful trials (size argument in stats::rnbinom)
(to sample library size)}

\item{swapClus}{a vector specifies two branches we want to swap their proportions
(use the root node labels for these two branches); By default,
NULL is set, which means we don't swap any two branches.}

\item{diffClus}{a vector specifies branches which we want to add signal
(differentially abundant) to. The signal is created by directly
changing their fold change; By default, NULL is set,
which means we don't change fold change of any branch.
(use the root node label for each branche)}

\item{FC}{a numeric vector indicates the fold change. The order of FC and
diffClus are matched. If diffClus is set to NULL, FC should be set to NULL too}
}
\value{
a matrix with rows representing tips of the tree and columns representing
samples. Samples in condition 1 and condition 2 are named corresponding with
the start string "C1_" and string "C2_"
}
\description{
\code{simuCount} could simulate a count table for all tips and nodes of a tree
on two different conditions. The signal on a specified part of a tree could be
easily simulated by changing arguments like swapClus or diffClus in \cpde{simuCount}
function. The proportions of all tips are estimated from the real data.
}
\examples{
{
library(dirmult)
library("GUniFrac")
data(throat.tree)
data(throat.otu.tab)

Wtree <- addNodeLab(treeO = throat.tree,
nodeLab = paste("Node",1:throat.tree$Nnode,sep=""))

DF <- simuCount(RealDat = throat.otu.tab,
wtree = Wtree,nClus = 40,nSam = 5,
muNB = 10000,sizeNB = 5,
swapClus = c("cluster1","cluster19"),
diffClus = NULL,FC=NULL)
}
}
