% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{square}
\alias{square}
\title{square a number}
\usage{
square(x)
}
\arguments{
\item{x}{Anumeric value to be squared}
}
\value{
The square of the input
}
\description{
Takes in any numeric value and squares it.
}
