% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{throat.otu.tab}
\alias{throat.otu.tab}
\title{OTU count table from 16S sequencing of the throat microbiome samples.}
\format{More details could be found in \code{\link[GUniFrac]{throat.otu.tab}}}
\usage{
throat.otu.tab
}
\description{
OTU count table from 16S sequencing of the throat microbiome samples.
}
\keyword{datasets}
