% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{throat.tree}
\alias{throat.tree}
\title{UPGMA tree of the OTUs from 16S sequencing of the throat microbiome samples.}
\format{More details could be found in \code{\link[GUniFrac]{throat.tree}}}
\usage{
throat.tree
}
\description{
UPGMA tree of the OTUs from 16S sequencing of the throat microbiome samples.
}
\keyword{datasets}
