% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tipComp.R
\name{tipComp}
\alias{tipComp}
\title{Compare the number of descendant tips in a result list to the truth}
\usage{
tipComp(rList, trueList, stree)
}
\arguments{
\item{rList}{a list includes all the inferred significant locations
(label of tips and nodes)}

\item{trueList}{a list includes all the true significant locations}
}
\value{
a data frame with 6 columns.

Inferred:  the number of tips under the inferred location (tip return 1).

Real: the true tip number under a real signal node

class:  the real class of inferred location (NoSignal / Cluster number)

NumTip:  tip number in the corresponding class

Distance: the shortest distance between point (Inferred, Real) to the line defined by function x-y=0; the column "scaleFreq" is derived from

scaleFreq: round((Freq*Inferred/NumTip))
}
\description{
After applying min P-value algorithm (using function find.p.min), a character vector
includes the labels of some tips and nodes is suggested to be the location with signals.
For each location (node or tip) in the result, we could use function tipComp to compare
it with the truth. A pair value (m,n) is returned, where m is the number of descendant
tips for the inferred location (tip return 1) and n is the true number of descendant
tips for the inferred location.
}
