% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topNodes.R
\name{topNodes}
\alias{topNodes}
\title{Table of the top differential nodes}
\usage{
topNodes(data, sort.by = "FDR", decreasing = FALSE,
  col.rowData = NULL, col.linkData = NULL, use.assays = NULL)
}
\arguments{
\item{data}{A tree summarizedExperiment object output from
\code{\link{treeAGG}}}

\item{sort.by}{The name of the column to sort by}

\item{decreasing}{A logical value. TRUE or FALSE. Should the sorting be
decreasing or not.}

\item{col.rowData}{The names of columns to be extracted from \code{rowData}.}

\item{col.linkData}{The names of columns to be extracted from \code{linkData}.}

\item{use.assays}{A numeric vector. It specifies which table to show. If
NULL, the table correspoinding to the first number storing in the
\code{use.assays} (in \code{metadata} of \code{input} data) will be shown.
To recall, numbers in the \code{use.assays} record which matrix-like elements
in the \code{assaysTable} have been used to do data analysis.}
}
\value{
a list of data frame
}
\description{
\code{topNodes} extracts the top differential nodes in a data frame or a list
of data frame for a given pair of group, ranked by p-value or absolute change
(e.g. log-fold change)
}
\author{
Ruizhu HUANG
}
