% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpr.R
\name{tpr0}
\alias{tpr0}
\title{Calculate the number of true positives and the total number of positives on
a tree structure}
\usage{
tpr0(tree, truth = NULL, found = NULL, level = c("node", "leaf"))
}
\arguments{
\item{tree}{A phylo object}

\item{truth}{Nodes that have signals (eg. differentally abundant at different
experimental conditions.).}

\item{found}{Nodes that have been found to have signal}

\item{level}{If "leaf", true positive rate is calculated at leaf level; if
"node", it is calculated at node level.}
}
\value{
a vector
}
\description{
\code{tpr0} calculates the number of true positives and the total number of
positives on a tree structure at leaf or node level.
}
\keyword{internal}
