% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transNode.R
\name{transNode}
\alias{transNode}
\title{Transfer between node number and node label}
\usage{
transNode(tree, input, use.original = FALSE, message = FALSE)
}
\arguments{
\item{tree}{A phylo object}

\item{input}{A character or numeric vector representing tree node label(s)
or tree node number(s)}

\item{use.original}{A logical value, TRUE or FALSE. Only required when
\strong{input} is numeric. Default is FALSE. There are cases that a tree
has no labels for some nodes. If keep.null is TRUE, NA is returned for these
nodes; otherwise, the node number (from \strong{input}) will be returned
with prefix "Node_" (for internal nodes) or "Leaf_" (for leaf nodes).}

\item{message}{A logical value, TRUE or FALSE. The default is FALSE. If TRUE,
message will show when a tree have duplicated labels for some internal nodes.}
}
\value{
a vector
}
\description{
\code{transNode} does the transformation between the number and the label of
a node on a tree
}
\examples{
library(ggtree)

data(tinyTree)

ggtree(tinyTree, branch.length = 'none') +
geom_text2(aes(label = label), hjust = -0.3) +
geom_text2(aes(label = node), vjust = -0.8,
hjust = -0.3, color = 'blue')

#check whether the node number and node label are matched
transNode(tinyTree, input = c(11, 2, 4))

transNode(tinyTree, input = c('Node_16', 'Node_11'))

}
\author{
Ruizhu Huang
}
