% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treePath.R
\name{tree.path}
\alias{tree.path}
\title{display all nodes & tips
show significant nodes & tips}
\usage{
tree.path(TabTN, P.lim, wtree, stree, nClus, cluster, ClusDiff, collapse,
  show.minP, show.all, RangeAll, varCir, LegendName, title, alpha = 0.9,
  size.minP, whole, pch.cir = c(1, 19), ...)
}
\description{
display all nodes & tips
show significant nodes & tips
}
