% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeLevel.R
\name{treeLevel}
\alias{treeLevel}
\title{convert tree structure to matrix structure}
\usage{
treeLevel(wtree, stree)
}
\arguments{
\item{wtree}{the entire tree (phylo class)}

\item{stree}{a list of subtrees, each of which is cut at a node of wtree}
}
\value{
a matrix. each tip occupy a row.

each row includes all nodes in the path connecting the corresponding tip
and the root. Columns from left to right includes tree level from high to low.
The first column the root label.
}
\description{
\code{treeLevel} convert a tree structure into a matrix structure with
tips in the rows and tree levels in columns.
}
\examples{
\\notrun{
library(ape)
n <- 20
Lab <- paste("Node",1:(n-1),sep="")
tipLab <- paste("T",1:n,sep="")
# entire tree
Tree <- rtree(n,tip.label= tipLab)
Tree <- addNodeLab(treeO = Tree, nodeLab = Lab)

# Subtrees
sTree <- pruneTree(Tree)


Mtree <- treeLevel(wtree = Tree, stree = sTree)
}
}
