% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treePlot.R
\name{treePlot}
\alias{treePlot}
\title{visualize the phylogenetic tree}
\usage{
treePlot(tree, branch = NULL, col.branch = "blue", col.other = "grey",
  point = NULL, col.point = "orange", size.point = 2, zoomNode = NULL,
  zoomLevel = NULL, zoomScale = NULL, legend.title = c(branch =
  "Title_branch", point = "Title_point"), size.line.legend = 2,
  size.point.legend = 2, ...)
}
\arguments{
\item{tree}{a phylo object}

\item{branch}{a vector of node numbers or node labels to specify the branches to be colored. Each branch is represented by its root node(node numbers or labels). A leaf node reprents the edge connecting the leaf and its parent.}

\item{col.branch}{a vector of colors. Its length should be one or equals to the length of \strong{branch}. If the lengths of col.branch and branch are the same, their corresponding orders are matched. The default is blue.}

\item{col.other}{a color for the branches other than those specified in branch}

\item{point}{a vector of node numbers or node labels to specify the point locations in the tree}

\item{col.point}{the color for the \strong{point}.}

\item{size.point}{the size for the \strong{point}.}

\item{zoomNode}{a vector of nodes to be zoomed in. If default (null), the tree is not zoomed in.}

\item{zoomLevel}{a numeric vector. Its length is equal to 1 or equal to the length of zoomNode. If default (null), a leaf is zoomed in its direct parent level and a internal node is zoomed in its own level.}

\item{zoomScale}{a scale to zoom in. If default (null), tree is not zoomed in.}

\item{legend.title}{a vector to specify the title of the legend. It must be named with "branch" and "point" to match with the argument \strong{branch} and \strong{point}.}

\item{size.line.legend}{the line size shown in the legend for \strong{branch}}

\item{size.point.legend}{the point size shown in the legend for \strong{point}.}

\item{...}{see also \code{\link[ggtree]{ggtree}}}
}
\value{
a tree plot
}
\description{
\code{treePlot} is to visualize a phylogenetic tree.
}
\details{
treePlot is created based on the \pkg{ggtree} and \pkg{ggplot2}. So other geoms from these two packages could be combined with \code{treePlot} to add geoms in the figure created by \code{treePlot}.
}
\examples{

data(exTree)
ggtree(bigTree, layout = "circular", branch.length = "none")

# If we want to color two branches with root node 1000 and 1400
p <- treePlot(tree = bigTree, branch = c(1000, 1400))
p

# To use different color
p <- treePlot(tree = bigTree, branch = c(1000, 1400),
col.branch = c("red", "blue"), col.other = c("other" = "grey20"))
p

# To add legend, we could do it in this way
# the name of col.other would be used as the label in the legend
p + ggplot2::theme(legend.position = "right")

# To change the legend title
p <- treePlot(tree = bigTree, branch = c(1000, 1400),
col.branch = c("red", "blue"), col.other = c("other" = "grey20"),
legend.title = c("branch" = "Truth"))

p + ggplot2::theme(legend.position = "right")

# To zoom in the branch with root node 1400 (by scale 4)
p <- treePlot(tree = bigTree, branch = c(1000, 1400),
col.branch = c("red", "blue"), col.other = c("other" = "grey20"),
legend.title = c("branch" = "Truth"), zoomNode = 1400,
zoomScale = 8)

p + ggplot2::theme(legend.position = "right")

# If its sister branch is also interested, we could zoom in their parent level by zoomLevel = 1

p <- treePlot(tree = bigTree, branch = c(1000, 1400),
col.branch = c("red", "blue"), col.other = c("other" = "grey20"),
legend.title = c("branch" = "Truth"), zoomNode = 1400,
zoomScale = 8, zoomLevel = 1)

p + ggplot2::theme(legend.position = "right")

# If we want to add some points at some nodes of the tree
p <- treePlot(tree = bigTree, branch = c(1000, 1400),
col.branch = c("red", "blue"), col.other = c("other" = "grey20"),
legend.title = c("branch" = "Truth"), zoomNode = 1400,
zoomScale = 8, zoomLevel = 1, point = c(50, 100, 500, 1200, 1500))

# To change the legend for the point
p <- treePlot(tree = bigTree, branch = c(1000, 1400),
col.branch = c("coral", "blue"), col.other = c("other" = "darkgrey"),
legend.title = c("branch" = "Truth", "point" = "Estimate"),
 zoomNode = 1400,
zoomScale = 8, zoomLevel = 1, point = c(50, 100, 500, 1200, 1500),
col.point = c("found" = "violet"))

p + ggplot2::theme(legend.position = "bottom",
legend.text = element_text(size= 10),
legend.key.size = unit(4,"cm"),
legend.key.height = unit(0.4,"cm"),
legend.key.width = unit(0.5, "cm"),
legend.title = element_text(size = 15))
}
