% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeSignal.R
\name{treeSignal}
\alias{treeSignal}
\title{label true signal and found signal in a tree (the legend part is weird, fix this later)}
\usage{
treeSignal(ModOutp, TipDiff, colSel, sizeSel, pchSel, alpha, wtree, stree,
  P.lim = 0.05, VarCol = "FDR", LegendName, ptree, ...)
}
\arguments{
\item{ModOutp}{a dataframe output from differenetial analysis package (such as: edgeR)}

\item{TipDiff}{differential abundant tips (charactor vector)}

\item{colSel}{3 colors selected to label signal location (true, found, matched)}

\item{sizeSel}{3 sizes selected to label signal location (true, found, matched)}

\item{pchSel}{3 point types selected to label signal location (true, found, matched)}

\item{alpha}{value for the color transparency of points}

\item{wtree}{the entire phylogenetic tree which is based to do the aggregation}

\item{stree}{a list of subtrees which are cut at the nodes of phylogenetic tree}

\item{P.lim}{threshold for adjusted p-value (significant level)}

\item{VarCol}{the column name of adjusted p-value}

\item{LegendName}{the title of legend}

\item{ptree}{the part of the tree to be plotted}
}
\value{
a tree
}
\description{
\code{treeSignal} returns a tree plot with true signal location and found signal
location labelled with different colors
}
\examples{
\\notrun{
load("/Volumes/fiona/simu_500/R package/TreeInfer/data/Data_calcRateP.RData")
load("/Volumes/fiona/simu_500/R package/TreeInfer/data/tree.RData")

edg1 <- edgeRes[[1]]

treeSignal (ModOutp=edg1, TipDiff=c(TipC15,TipC19),
colSel=c("blue","firebrick","darkorchid")
,pchSel = c(1,1,1),
alpha=1,wtree=Wtree, stree=Stree, sizeSel = c(2,1,2),
LegendName=" ",P.lim = 0.05,
VarCol = "FDR",ptree=wtree,size=0.1)
}
}
