% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classAccessors.R
\docType{methods}
\name{colData}
\alias{colData}
\alias{colData,treeSummarizedExperiment-method}
\title{Accessor functions for treeSummarizedExperiment class}
\usage{
colData(x)

\S4method{colData}{treeSummarizedExperiment}(x)
}
\arguments{
\item{x}{A treeSummarizedExperiment object}
}
\description{
Accessor functions for treeSummarizedExperiment class
}
