% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClass.R
\docType{class}
\name{treeSummarizedExperiment-class}
\alias{treeSummarizedExperiment-class}
\title{An S4 class treeSummarizedExperiment}
\description{
The class \code{treeSummarizedExperiment} is extended from the class
\code{\link[SummarizedExperiment]{SummarizedExperiment}} by including a
\strong{phylo} object in the metadata. It further restricts the rowData to
have a column named \strong{nodeLab}. This column provides the labels of
nodes on the tree that are corresponding to the rows of matrix in
\strong{assays}. If multiple rows of a matrix in \strong{assays} are matched
to a same node on the tree, then the same node label is shared by these rows.
}
