% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClass.R
\name{treeSummarizedExperiment-constructor}
\alias{treeSummarizedExperiment-constructor}
\alias{treeSummarizedExperiment}
\title{A Constructor for treeSummarizedExperiment}
\usage{
treeSummarizedExperiment(tree = NULL, linkData = NULL, ...)
}
\arguments{
\item{tree}{A phylo object}

\item{linkData}{A data frame or \code{\link[S4Vectors]{DataFrame}}. It has
the same number of rows as the matrix-like element in \code{assays.}. The
row order between \code{linkData} and the matrix-like element in
\code{assays.} is the same. It includes columns as below.
\itemize{
\item \strong{nodeLab} the labels of nodes
\item \strong{nodeLab_alias} a alias of column \code{nodeLab}. It is
created only when there are missing value or duplicated value in column
\code{nodeLab}. Commonly, it is a prefix "Node_" follow by the value in
column \code{nodeNum}.
\item \strong{nodeNum} the numbers of nodes
\item \strong{isLeaf} is it a leaf node
\item \strong{rowID} the row number in the matrix-like element in
\code{assays} }}

\item{...}{see arguments in
\code{\link[SummarizedExperiment]{RangedSummarizedExperiment-class}}
\itemize{
\item{\strong{assays}:}{ a list of matrix-like elements, or a matrix-like
object. All elements of the list must have the same dimensions, and
dimension names (if present) must be consistent across elements and with
the row names of \strong{rowData} and \strong{colData}}.

\item{\strong{rowData}:}{ A \code{\link[S4Vectors]{DataFrame-class}} object
describing the rows. Row names, if present, become the row names of the
\strong{leafSummarizedExperiment}. The number of rows of the
\code{\link[S4Vectors]{DataFrame-class}} must equal the number of rows of the
matrices in \strong{assays}. }

\item{\strong{colData}:}{ An optional
\code{\link[S4Vectors]{DataFrame-class}} describing the samples. Row names,
if present, become the column names of the
\strong{leafSummarizedExperiment}.}

\item{\strong{metadata}:}{ An optional list of arbitrary content describing
the overall experiment.}
}}
}
\value{
a treeSummarizedExperiment object
}
\description{
\code{treeSummarizedExperiment} constructs a treeSummarizedExperiment.
}
\seealso{
\code{\link{treeSummarizedExperiment}}
  \code{\link{treeSummarizedExperiment-accessor}}
  \code{\link{leafSummarizedExperiment}}
  \code{\link[SummarizedExperiment]{SummarizedExperiment-class}}
}
\author{
Ruizhu HUANG
}
