% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classBuilder.R
\name{treeSummarizedExperiment}
\alias{treeSummarizedExperiment}
\title{create a treeSummarizedExperiment object}
\usage{
treeSummarizedExperiment(tree, ...)
}
\arguments{
\item{tree}{A phylo object.}

\item{...}{see arguments for
\code{\link[SummarizedExperiment]{SummarizedExperiment}}
\itemize{
\item{\strong{assays}:}{ a list of matrix-like elements, or a matrix-like
object. All elements of the list must have the same dimensions, and
dimension names (if present) must be consistent across elements and with
the row names of \strong{rowData} and \strong{colData}}.
\item{\strong{rowData}:}{ A \code{\link[S4Vectors]{DataFrame-class}}
describing the rows. Row names, if present, become the row names of the
\strong{treeSummarizedExperiment}. The number of rows of the DataFrame must
equal the number of rows of the matrices in \strong{assays}. Note: a column
named \strong{nodeLab} is required to provide the corresponding nodes on the
tree for each row of the matrices in \strong{assays}.}
\item{\strong{colData}:}{ An optional DataFrame describing the samples.
Row names, if present, become the column names of the
\strong{treeSummarizedExperiment}.}
\item{\strong{metadata}:}{ An optional list of arbitrary content describing the overall
 experiment.}}}
}
\value{
A treeSummarizedExperiment object
}
\description{
\code{treeSummarizedExperiment} is to create a object with
\strong{treeSummarizedExperiment} class. The class
\strong{treeSummarizedExperiment} is an extension to
\strong{SummarizedExperiment}.
}
\examples{
# tree
data("tinyTree")

# assays
count <- matrix(rpois(190, 50), nrow = 19)
rownames(count) <- c(tinyTree$tip.label, tinyTree$node.label)
colnames(count) <- paste("C_", 1:10, sep = "_")

# colData
sampC <- data.frame(condition = rep(c("control", "trt"), each = 5),
gender = sample(x = 1:2, size = 10, replace = TRUE))
rownames(sampC) <- colnames(count)

tse <- treeSummarizedExperiment(tree = tinyTree, assays = list(count),
colData = sampC)

}
\seealso{
\code{\link[SummarizedExperiment]{SummarizedExperiment}}
}
