% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateContrast.R
\name{updateContrast}
\alias{updateContrast}
\title{update the value of contrast in treeSummarizedExperiment object}
\usage{
updateContrast(data, contrast)
}
\arguments{
\item{data}{A treeSummarizedExperiment object. The output from
\code{\link{runEdgeR}}.}

\item{contrast}{A numeric vector to define contrast.}
}
\value{
A treeSummarizedExperiment object. The overall structure is the same
  to the output of \code{\link{runEdgeR}}. The updates occur in the
  \code{assays} and the \code{contrast} (one element of \code{metadata}).
\item{assays}{It stores a list of tables extracted from a \code{DGELRT}
object that is generated by \code{\link[edgeR]{glmLRT}} with the specified
\code{contrast}}
\item{rowData}{It stores the information of rows in \code{assays} as input}
\item{colData}{NULL}
\item{metadata}{
   \itemize{
   \item \code{assaysInput} a list of matrix-like elements. Stay the same as
   the \code{assaysInput} in input \code{data}.
   \item \code{use.assays} a numeric vector. Stay the same as the
   \code{use.assays} in input \code{data}.
   \item \code{design} a design matrix. Stay the same as the \code{design} in
   input \code{data}.
   \item \code{contrast} the contrast vector as input.
   \item \code{dgeGLM} a list of \code{DGEGLM} objects. Stay the same as the
   \code{dgeGLM} in input \code{data}.
   }
}
}
\description{
\code{updateContrast} updates the output of \code{\link{runEdgeR}} when a
different contrast is specified. It conducts likelihood ratio test for a
given coefficient contrast by using the function \code{\link[edgeR]{glmLRT}}.
}
\seealso{
\code{\link{runEdgeR}} \code{\link[edgeR]{glmFit}}
  \code{\link[edgeR]{glmLRT}} \code{\link[edgeR]{DGEGLM-class}}
}
\author{
Ruizhu HUANG
}
