% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateTSE.R
\name{updateTSE}
\alias{updateTSE}
\title{replace assays in treeSummarizedExperiment}
\usage{
updateTSE(assays, tse)
}
\arguments{
\item{assays}{A list of matrix-like elements.  It should have the same number
of matrix-like elements as the \code{assays} of data specified in the
argument \code{tse}. Also, its matrix-like elements should have  exactly
the same number of rows and rows representing the same thing as those in
the old \code{assays} of \code{tse}. However, the columns could be
different.}

\item{tse}{A treeSummarizedExperiment object.}
}
\value{
A treeSummarizedExperiment object.
}
\description{
\code{updateTSE} allows to replaces the matrix-like element in assays of
\code{treeSummarizedExperiment} with matrix-like element that have the same
number of rows but different number of columns.
}
\details{
The new \code{assays} have exactly the same rows as the original one that was
replaced. But the columns are allowed to be different. The \code{colData} in
the original \code{treeSummarizedExperiment} object is removed from the new
one.
}
\author{
Ruizhu HUANG
}
