% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classValid.R
\name{validTSE}
\alias{validTSE}
\title{valid treeSummarizedExperiment class}
\usage{
validTSE(obj)
}
\arguments{
\item{obj}{A treeSummarizedExperiment object}
}
\value{
TRUE or a character string.
}
\description{
\code{validTSE} is to valid treeSummarizedExperiment object.
}
\keyword{internal}
