% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treePlot.R
\name{addLegend}
\alias{addLegend}
\title{Add legend
\code{addLegend} customizes the legend.}
\usage{
addLegend(legend.theme)
}
\arguments{
\item{legend.theme}{A list of arguments used for the theme in ggplot2 package (see \code{\link[ggplot2]{theme}} ) and starting with "legend."}
}
\description{
Add legend
\code{addLegend} customizes the legend.
}
