% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treePlot.R
\name{addLegend}
\alias{addLegend}
\title{Add legend
\code{addLegend} customizes the legend.}
\usage{
addLegend(legend.theme = list(NULL))
}
\arguments{
\item{legend.theme}{A list of arguments used for the theme in ggplot2 package
(see \code{\link[ggplot2]{theme}} ) and starting with "legend."}
}
\value{
a list
}
\description{
Add legend
\code{addLegend} customizes the legend.
}
\author{
Ruizhu Huang
}
\keyword{internal}
