% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addRank.R
\name{addRank}
\alias{addRank}
\title{create rank variable}
\usage{
addRank(ModOutp, varEst, varSE, ...)
}
\arguments{
\item{ModOutp}{a dataframe output from differenetial analysis package (such as: edgeR)}

\item{varEst}{the name of the column which contains the main effect estimate}

\item{varSE}{the name of the column which contains the standard error of main effect estimate}

\item{...}{see the arguments except data in \code{rvalues::rvalues}}
}
\value{
a dataframe
This function returns a new data frame which is actually the input dataframe ModOutp + a
new created column rvalue.

rvalue is r value
}
\description{
\code{addRank} is to create a rank variable which could be used later in
the tree aggregateion.
}
\examples{
{

library(DESeq2)
# generate count tables from RNA-Seq data
 cnts <- matrix(rnbinom(n=1000, mu=100, size=1/0.5), ncol=10)
 rownames(cnts) <- 1:nrow(cnts)
 cond <- factor(rep(1:2, each=5))

# object construction
dds <- DESeqDataSetFromMatrix(cnts, DataFrame(cond), ~ cond)

# standard analysis
dds <- DESeq(dds)
res <- results(dds)

# compute r-value
Rres <- addRank(ModOutp = res, varEst = "log2FoldChange", varSE = "lfcSE",
varRank = "rvalue")
#' }
}
