% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branchPlot.R
\name{branchPlot}
\alias{branchPlot}
\title{Plot a tree branch}
\usage{
branchPlot(branch, CladeLab, CladeCol, nodeLab = NULL, LegendName)
}
\arguments{
\item{branch}{a selected branch to be plotted (phylo object)}

\item{CladeLab}{a vector includes the labels of root node for interested clades
and "other" (label for the branch part which are not included in the interested
clades)}

\item{CladeCol}{a vector of colors to be assigned to the clades in CladeLab. The order between
CladeLab and CladeCol is matched}

\item{nodeLab}{a vector. Labels of nodes to be annotated in the plot; default is NULL,
which indicates no node is annotated in the plot}

\item{LegendName}{the name of the legend title.}
}
\value{
a plot
}
\description{
\code{branchPlot} could show a tree plot of a selected branch with annotated
nodes and clades
}
