% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clusters.R
\name{ClusterInf}
\alias{ClusterInf}
\title{Information of clusters
\code{ClusterInf} summarize the total number of tips in the cluster
and the proportion of counts for each cluster calculated from the real data}
\usage{
ClusterInf(RealDat, wtree, nClus, stree)
}
\arguments{
\item{RealDat}{count table (include at least all tips of the tree)}

\item{nClus}{number of clusters}

\item{tree}{a phylo tree to be cut;}
}
\value{
a data frame

First column indicates the cluster name
Second column indicates the total number of tips in the cluster
Third column indicates the proportions of counts for the clusters which are estimated
from the real data use \code{dirmult} from package dirmult
}
\description{
Information of clusters
\code{ClusterInf} summarize the total number of tips in the cluster
and the proportion of counts for each cluster calculated from the real data
}
