% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doTable.R
\name{doFC}
\alias{doFC}
\title{generate the fold change}
\usage{
doFC(scene, BranchA, BranchB, tree, ratio, data, pct)
}
\arguments{
\item{scene}{Scenarios (\dQuote{S1}, \dQuote{S2}, \dQuote{S3})}

\item{BranchA}{the branch node label of branch A.}

\item{BranchB}{the branh node label of branch B.}

\item{tree}{a phylo object}

\item{ratio}{the proportion ratio between \code{BranchB} and \code{BranchA} (B/A)}

\item{data}{the real data (count table)}

\item{pct}{the percentage (in number) of the leaves in BranchA that will swap with BranchB.}
}
\value{
a numeric vector
}
\description{
\code{doFC} is to generate a fold change for different scenarios
}
