% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doTable.R
\name{doTable}
\alias{doTable}
\title{simulate a count table}
\usage{
doTable(tree, data, scene = "S1", from.A = NULL, from.B = NULL,
  minTip.A = 0, maxTip.A = Inf, minTip.B = 0, maxTip.B = Inf,
  minPr.A = 0, maxPr.A = 1, ratio = 2, pct = 0.6, nSam = c(50, 50),
  mu = 50, size = 10000, n = 1, fun = sum)
}
\arguments{
\item{tree}{a phylo object}

\item{data}{a matrix. A count table from real data. It has the entities corresponding to tree leaves in the row and samples in the column.}

\item{scene}{\dQuote{S1}, \dQuote{S2}, or \dQuote{S3} (see \bold{Details}). Default is \dQuote{S1}.}

\item{from.A, from.B}{the branch node label of branch A,B. Default, both are NULL. In simulation, we select two branches (A & B) to have differential abundance under different conditions. One could specify these two branches or let \code{doTable} choose. (Note: If \code{from.A} is NULL, \code{from.B} is set to NULL).}

\item{minTip.A}{the minimum number of leaves in branch A}

\item{maxTip.A}{the maximum number of leaves in branch A}

\item{minTip.B}{the minimum number of leaves in branch B}

\item{maxTip.B}{the maximum number of leaves in branch B}

\item{minPr.A}{the minimum abundance proportion of leaves in branch A}

\item{maxPr.A}{the maximum abundance proportion of leaves in branch A}

\item{ratio}{the proportion ratio of branch B to branch A. This value is used to select branches(see \bold{Details}). If there are not branches exactly having this ratio, the pair with the value closest to \code{ratio} would be selected.}

\item{pct}{the percentage of leaves in branch B has the differential abundance under different conditions (only for scenario \dQuote{S3})}

\item{nSam}{a numeric vector with 2 length. the sample size for two different conditons}

\item{mu, size}{the parameters of negative binomial distribution. (see mu and size in \code{\link[stats]{rnbinom}}). Parameters used to generate library size for each sample.}

\item{n}{a numeric value to specify how many count tables would be generated with the same setting. Default is one and one count table would be obtained at the end. If above one, the output of \code{doTable} is a list of matrices (count tables). This is useful, when one needs multiple simulations.}

\item{fun}{a function to derive the count at each internal node based on its descendant leaves, e.g. sum, mean. The argument of the function is a numeric vector including the counts of an internal node's descendant leaves .}
}
\description{
\code{doTable} is to create a count table for all nodes of a tree under two different groups such that the tree would have different abundance patterns under different conditions.
}
\details{
\code{doTable} is to simulate a count table for entities which are corresponding to the nodes of a tree. The entities are in rows and the samples from different groups or conditions are in columns. The library size of each sample is sampled from a negative binomial distribution with mean and size specified by the arguments \code{mu} and \code{size} above. The counts of entities, which are located on the tree leaves, in the same sample are assumed to follow dirichlet-multinomial distribution. The parameters for the dirichlet-multinomial distribution are estimated from a real data specified by the argument \code{data} via the function \code{dirmult}(see \code{\link[dirmult]{dirmult}}). To generate different abundance patterns under different conditions, we provide three different scenarios, \dQuote{S1}, \dQuote{S2}, and \dQuote{S3} (specified via \code{scene}).
\itemize{
\item S1: two branches are selected to swap their proportions, and leaves on the same branch have the same fold change.
\item S2: two branches are selected to swap their proportions. Leaves in the same branch have different fold changes but same direction (either ncrease or decrease).
\item S3: two branches are selected. One branch has its proportion swapped with the proportion of some leaves from the other branch.
}
\code{doTable} could select the two branches automatically based on the criteria specified by \code{minTip.A}, \code{maxTip.A}, \code{minPr.A}, \code{maxPr.A}, \code{minTip.B}, \code{maxTip.B} and \code{ratio}. or find one branch when the other is fixed via \code{from.A}.
}
\examples{
{
if(require(GUniFrac)){
data("throat.otu.tab")
data("throat.tree")

dat <- doTable(tree = throat.tree,
data = as.matrix(t(throat.otu.tab)),
ratio = 2)
}



}
}
