% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exTree}
\alias{exTree}
\title{A simulated phylogenetic tree with 50 tips and 49 internal nodes}
\format{A phylo object with 10 tips and 9 internal nodes:
\describe{
  \item{Tip labels}{t1, t2, ..., t50.}
  \item{Node labels}{NULL}
}}
\usage{
exTree
}
\description{
A phlo object randomly created by using the function \code{\link[ape]{rtree}}
}
\keyword{datasets}
