% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdr.R
\name{fdr}
\alias{fdr}
\title{Calculate false discovery rate (fdr) on a tree structure}
\usage{
fdr(tree, truth, found, level = c("node", "leaf"), direction = FALSE)
}
\arguments{
\item{tree}{A phylo object}

\item{truth}{Nodes that have signals (eg. differentally abundant at different experimental conditions.). If the signals are in different directions (up and down), then provide the nodes as a list of two members (one up and one down). One could provide either node number or node label. \strong{Note:} When fdr at node level is required and only leaf nodes with signal are provided, then the internal nodes which are shared and only shared by the provied leaf nodes (signal in the same direction) will be found out and used with the leaf nodes in the fdr calculation; when fdr at leaf level is required and the given nodes have internal nodes, then the descendant leaf nodes will be found out and used in the fdr calculation.}

\item{found}{Nodes that have been found to have signal (eg. differentally abundant at different experimental conditions). If the signals are in different directions (up and down), then provide the nodes as a list of two members (one up and one down). One could provide either node number or node label. \strong{Note:} When fdr at node level is required, then the descendant nodes of the provied nodes (include themselves) will be found out and used in the fdr calculation; when fdr at leaf level is required, then the descendant leaf nodes will be found out and used in the fdr calculation.}

\item{level}{If "leaf", false discovery rate is calculated at leaf level; if "node", it is calculated at node level.}

\item{direction}{TRUE or FALSE. Default is FALSE. If TRUE, the signal direction is taken into account; the argument \strong{truth} and \strong{found} should both be a list of two members and the order of directions should match.}
}
\value{
A false discovery rate
}
\description{
\code{fdr} calculates the false discovery rate (fdr) on a tree structure at the leaf or node level.
}
\examples{

library(ggtree)
data("tinyTree")
ggtree(tinyTree) + geom_text2(aes(label = node))

# Truth: two branches have differential
# abundance under different conditions.
# branch with node 16 (direction: increase)
# branch with node 13 (direction: decrease)


# ------------ ignore direction -------------------
 # Found: branches with node 17 and node 14
 # fdr at the tip level
fdr1 <- fdr(tree = tinyTree, truth = c(16, 13),
            found = c(17, 14), level = "leaf")
 # fdr at the node level
fdr2 <- fdr(tree = tinyTree, truth = c(16, 13),
            found = c(17, 14), level = "node")


# ------------ direction matters-------------------
# Found: branches with node 17 (up) and node 14 (down)

 # the order of direction for truth and found should be the same
(fdr3 <- fdr(tree = tinyTree, truth = list(16, 13),
            found = list(16, 13), level = "leaf",
            direction = TRUE))  # correct order

(fdr4 <- fdr(tree = tinyTree, truth = list(16, 13),
            found = list(13, 16), level = "node",
            direction = TRUE)) # wrong order



}
