% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdr.R
\name{fdr0}
\alias{fdr0}
\title{Calculate false discoveries and discoveries}
\usage{
fdr0(tree, truth = NULL, found = NULL, only.Tip = TRUE)
}
\arguments{
\item{tree}{A phylo object}

\item{truth}{Nodes that have signals (eg. differentally abundant at different
experimental conditions).}

\item{found}{Nodes that have been found to have signal}

\item{only.Tip}{A logical value, TRUE or FALSE. If TRUE, false discovery rate
is calculated at the leaf (tip) level; otherwise it is calculated at node
level. The default is TRUE}
}
\value{
a vector
}
\description{
\code{fdr0} calculates the number of false discoveries and the total number
of discoveries at leaf or node level on a tree structure .
}
\author{
Ruizhu Huang
}
\keyword{internal}
