% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findAncestor.R
\name{findAncestor}
\alias{findAncestor}
\title{find the ancestors of specified nodes}
\usage{
findAncestor(tree, node, level, treeMat = NULL)
}
\arguments{
\item{tree}{a phylo object}

\item{node}{a vector of node numbers or node labels}

\item{level}{a vector of numbers to define n generation before the specified nodes}

\item{treeMat}{a matrix with each row representing a path. The entry is node number. The first column is leave node numbers. The columns are organized as the order of nodes in the paths connecting leaves and the root. Default is null, treeMat is generated automatically. The treeMat is provided when the tree is very large to save running time.}
}
\value{
a vector of node numbers
}
\description{
\code{findAncestor} is to find the ancestor in the n generation before specified nodes.
}
\examples{
library(ggtree)
data(exTree)
ggtree(exTree, branch.length = 'none') +
 geom_text2(aes(label = node))

 findAncestor(tree = exTree, node = c(53, 61), level = 1)
}
