% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findOS.R
\name{findOS}
\alias{findOS}
\title{find descendants (or offsprings)}
\usage{
findOS(ancestor, tree, only.Tip = TRUE, self.include = TRUE)
}
\arguments{
\item{ancestor}{an internal node. It could be the node number or the node label.}

\item{tree}{a phylo object.}

\item{only.Tip}{a logical value, TRUE or FALSE. The default is TRUE. If default, only the leaf nodes in the descendant nodes would be returned.}

\item{self.include}{a logical value, TRUE or FALSE. The default is TRUE. If default, the node specified in \strong{ancestor} is included. The leaf node itself is returned as its descendant.}
}
\description{
\code{findOS} is to find descendants of an internal node.
}
\examples{
data(tinyTree)

library(ggtree)
ggtree(tinyTree) +
geom_text2(aes(label = node)) +
geom_hilight(node = 15, fill = 'steelblue', alpha = 0.5)

(tips <- findOS(15, tinyTree, only.Tip = TRUE))

}
