% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewSim.R
\name{findParallel}
\alias{findParallel}
\title{find parallel nodes}
\usage{
findParallel(tree, input, label = FALSE)
}
\arguments{
\item{tree}{A phylo object.}

\item{input}{A numeric or character vector. Node labels or node numbers.}

\item{label}{A logical value, TRUE or FALSE. If TRUE, node labels are
returned.}
}
\value{
A vector of node labels or node numbers
}
\description{
\code{findParallel} is to find target nodes that could build up a tree with
some specified nodes. The returned target nodes is the combination that has
the minimum number of nodes. In other words, a tree is cut in a way that
the number of branches is the minimum in all possibilities and the specified
branches are obtained. A branch is represented by its branch node. A leaf
node represents the edge connecting the leaf and its parent.
}
\keyword{internal}
