% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findSibling.R
\name{findSibling}
\alias{findSibling}
\title{find the sibling node}
\usage{
findSibling(tree, input, label = FALSE)
}
\arguments{
\item{tree}{A phylo object.}

\item{input}{A numeric or character vector. Node labels or node numbers.}

\item{label}{A logical value, TRUE or FALSE. If TRUE, node labels are
returned.}
}
\value{
A vector of node labels or node numbers
}
\description{
\code{findSibling} is to find the sibling node of an \code{input} node.
}
\keyword{internal}
