% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClass.R
\name{leafSummarizedExperiment-constructor}
\alias{leafSummarizedExperiment-constructor}
\alias{leafSummarizedExperiment}
\title{constructor for leafSummarizedExperiment object}
\usage{
leafSummarizedExperiment(tree, ...)
}
\arguments{
\item{tree}{A phylo object.}

\item{...}{see arguments in
\code{\link[SummarizedExperiment]{RangedSummarizedExperiment-class}}
\itemize{
\item{\strong{assays}:}{ a list of matrix-like elements, or a matrix-like
object. All elements of the list must have the same dimensions, and
dimension names (if present) must be consistent across elements and with
the row names of \strong{rowData} and \strong{colData}}.

\item{\strong{rowData}:}{ A \code{\link[S4Vectors]{DataFrame-class}} object
describing the rows. Row names, if present, become the row names of the
\strong{leafSummarizedExperiment}. The number of rows of the
\code{\link[S4Vectors]{DataFrame-class}} must equal the number of rows of the
matrices in \strong{assays}. Note: a column named \strong{nodeLab} or
rownames, which provides the labels of corresponding nodes on the tree for
each row of the matrices in \strong{assays}, is required.}

\item{\strong{colData}:}{ An optional
\code{\link[S4Vectors]{DataFrame-class}} describing the samples. Row names,
if present, become the column names of the
\strong{leafSummarizedExperiment}.}

\item{\strong{metadata}:}{ An optional list of arbitrary content describing
the overall experiment.}
}}
}
\value{
A leafSummarizedExperiment object
}
\description{
\code{leafSummarizedExperiment} is to create a
\strong{leafSummarizedExperiment} object.
}
\examples{
# tree
data("tinyTree")

# assays
count <- matrix(rpois(190, 50), nrow = 19)
rownames(count) <- c(tinyTree$tip.label, tinyTree$node.label)
colnames(count) <- paste("C_", 1:10, sep = "_")

# colData
sampC <- data.frame(condition = rep(c("control", "trt"), each = 5),
gender = sample(x = 1:2, size = 10, replace = TRUE))
rownames(sampC) <- colnames(count)

tse <- leafSummarizedExperiment(tree = tinyTree, assays = list(count),
colData = sampC)

}
\seealso{
\code{\link{leafSummarizedExperiment}}
}
\author{
Ruizhu HUANG
}
