% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClass.R
\name{leafSummarizedExperiment-constructor}
\alias{leafSummarizedExperiment-constructor}
\alias{leafSummarizedExperiment}
\title{Construct a leafSummarizedExperiment object}
\usage{
leafSummarizedExperiment(tree, ...)
}
\arguments{
\item{tree}{A phylo object.}

\item{...}{see the arguments for the constructor in
\code{\link[SummarizedExperiment]{RangedSummarizedExperiment-class}}
\itemize{
\item{\strong{assays}:}{ a list of matrix-like elements, or a matrix-like
object. All elements of the list must have the same dimensions, and
dimension names (if present) must be consistent across elements and with
the row names of \strong{rowData} and \strong{colData}}.

\item{\strong{rowData}:}{ A \code{\link[S4Vectors]{DataFrame-class}} object
describing the rows. Row names, if present, become the row names of the
\strong{leafSummarizedExperiment}. The number of rows of the
\code{\link[S4Vectors]{DataFrame-class}} must equal the number of rows of the
matrices in \strong{assays}. Note: a column named \strong{nodeLab} or
rownames, which provides the labels of corresponding nodes on the tree for
each row of the matrices in \strong{assays}, is required.}

\item{\strong{colData}:}{ An optional argument with
\code{\link[S4Vectors]{DataFrame-class}} describing the samples. Row names,
if present, become the column names of the
\strong{leafSummarizedExperiment}.}

\item{\strong{metadata}:}{ An optional list of arbitrary content describing
the overall experiment.}
}}
}
\value{
A leafSummarizedExperiment object
}
\description{
\code{leafSummarizedExperiment} is the constructor of the
\strong{leafSummarizedExperiment} class.
}
\examples{
# tree
data("tinyTree")

# the count table
count <- matrix(rpois(100, 50), nrow = 10)
rownames(count) <- c(tinyTree$tip.label)
colnames(count) <- paste("C_", 1:10, sep = "_")

# the sample information
sampC <- data.frame(condition = rep(c("control", "trt"),
                                    each = 5),
                    gender = sample(x = 1:2, size = 10,
                                    replace = TRUE))
rownames(sampC) <- colnames(count)

# build a leafSummarizedExperiment object
lse <- leafSummarizedExperiment(tree = tinyTree,
                                assays = list(count),
                                colData = sampC)

}
\seealso{
\code{\link{leafSummarizedExperiment-class}}
  \code{\link{treeSummarizedExperiment-class}}
  \code{\link[SummarizedExperiment]{SummarizedExperiment-class}}
}
\author{
Ruizhu HUANG
}
