% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodeCount.R
\name{nodeCount}
\alias{nodeCount}
\title{Calculate the counts at internal nodes}
\usage{
nodeCount(data, tree, fun = sum)
}
\arguments{
\item{data}{A matrix or data frame. A count table from real data.}

\item{tree}{A phylo object}

\item{fun}{A function to create the count of an internal node based on the counts at its descendant leaf nodes. The default is sum}
}
\value{
A count table (matrix class) with a row representing a node and
a column representing a sample.
}
\description{
\code{nodeCount} calculates the counts at internal nodes. The count of an internal node is the sum of counts at its descendant leaves.
}
\examples{
data("tinyTree")
library(ggtree)
(p <- ggtree(tinyTree) + geom_text(aes(label = label)))

count <- matrix(rpois(100, 10), nrow =10)
rownames(count) <- tinyTree$tip.label
colnames(count) <- paste(c("C1_", "C2_"),
c(1:5, 1:5), sep = "")

count_tinyTree <- nodeCount(data = count,
tree = tinyTree, fun = mean)

# check to see whether the count of an internal node is the sum
# of counts of its descendant leaves.
# here, check the first sample as an example

nod <- transNode(tree = tinyTree, input = rownames(count_tinyTree))
d <- cbind.data.frame(node = nod, count = count_tinyTree[, 1])

ggtree(tinyTree) \%<+\% d + geom_text2(aes(label = count))
}
