% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodeValue.R
\docType{methods}
\name{nodeValue}
\alias{nodeValue}
\alias{nodeValue,phylo,matrixDataframe,function-method}
\alias{nodeValue,ANY,treeSummarizedExperiment,function-method}
\title{Calculate entity values at internal nodes}
\usage{
nodeValue(tree, data, fun)

\S4method{nodeValue}{phylo,matrixDataframe,`function`}(tree, data, fun = sum)

\S4method{nodeValue}{ANY,treeSummarizedExperiment,`function`}(tree, data,
  fun = sum)
}
\arguments{
\item{tree}{A phylo object to provide information of tree structure. We don't
need this if data is a treeSummarizedExperiment object.}

\item{data}{A data frame, matrix or treeSummarizedExperiment object. If it's
treeSummarizedExperiment, then tree information is also given.}

\item{fun}{A function to create the value of an internal node based on values
at its descendant leaf nodes. The default is sum.}
}
\value{
A treeSummarizedExperiment if input data is a
treeSummarizedExperiment object. A matrix if input data is a matrix or
data frame.
}
\description{
\code{nodeValue} calculates value, e.g., count, for each internal
node based on the values of its descendant leaves. For example, if the
abundance of a cell cluster, which corresponds to an internal node in the
tree, is of interest, we could sum the abundance of all cell types in that
cluster. Each cell type, in this case, corresponds to a leaf node in the
tree. Different calculations, instead of only sum, might be required. This
could be achieved by specifying different functions in the argument
\strong{fun}.
}
\examples{
data("tinyTree")
if(require(ggtree)){
(p <- ggtree(tinyTree) + geom_text(aes(label = label)))

count <- matrix(rpois(100, 10), nrow =10)
rownames(count) <- tinyTree$tip.label
colnames(count) <- paste(c("C1_", "C2_"),
c(1:5, 1:5), sep = "")

count_tinyTree <- nodeValue(data = count,
tree = tinyTree, fun = mean)

# check to see whether the count of an internal node is the sum
# of counts of its descendant leaves.
# here, check the first sample as an example

nod <- transNode(tree = tinyTree, input = rownames(count_tinyTree))
d <- cbind.data.frame(node = nod, count = count_tinyTree[, 1])

ggtree(tinyTree) \%<+\% d + geom_text2(aes(label = count))
}

}
\seealso{
\code{\link{treeSummarizedExperiment}} for treeSummarizedExperiment
creation.
}
\author{
Ruizhu Huang
}
