% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parEstimate.R
\docType{methods}
\name{parEstimate}
\alias{parEstimate}
\alias{parEstimate,matrix-method}
\alias{parEstimate,list-method}
\alias{parEstimate,leafSummarizedExperiment-method}
\title{Estimate parameters for Dirichlet distribution}
\usage{
parEstimate(data)

\S4method{parEstimate}{matrix}(data)

\S4method{parEstimate}{list}(data)

\S4method{parEstimate}{leafSummarizedExperiment}(data)
}
\arguments{
\item{data}{A count table. Samples in the column and entities in the row.}
}
\value{
A list including \dQuote{pi} and \dQuote{theta}
}
\description{
\code{parEstimate} estimates parameters \eqn{\pi} and \eqn{\theta} from
a real data (count table). The entity counts within each sample is assumed to
follow a multinomial distribution with probabilities,
\eqn{ \pi_{1}, \pi_{2}, ..., \pi_{k-1}}. When \eqn{\pi} follows a Dirichlet
distribution,  the marginal distribution of X, the entity count, is
Dirichlet-multinomial with parameters \eqn{\pi} and \eqn{\theta}.
More details see \code{\link[dirmult]{dirmult}}
}
\details{
\code{parEstimate} is created based on
\code{\link[dirmult]{dirmult}}).
}
\examples{
\dontrun{
library(GUniFrac)
data("throat.otu.tab")
throat.par <- parEstimate(data = t(throat.otu.tab))
}

}
\author{
Ruizhu Huang
}
