% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parEstimate.R
\name{parEstimate}
\alias{parEstimate}
\title{Estimate parameters for Dirichlet distribution}
\usage{
parEstimate(data)
}
\arguments{
\item{data}{A count table. Samples in the column and entities in the row.}
}
\value{
A list including \dQuote{pi} and \dQuote{theta}
}
\description{
\code{parEstimate} estimates parameters for a Dirichlet distribution from
a real data (count table).
}
\details{
Use the default setting from \code{dirmult}
(see \code{\link[dirmult]{dirmult}})
}
\examples{
library(GUniFrac)
data("throat.otu.tab")
throat.par <- parEstimate(data = t(throat.otu.tab))
}
\author{
Ruizhu Huang
}
