% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtree.R
\name{pruneTree}
\alias{pruneTree}
\title{prune tree at each node}
\usage{
pruneTree(wtree)
}
\arguments{
\item{wtree}{an object of phylo class}
}
\value{
a list of phylo object
}
\description{
This function prune the tree at each node.
}
\examples{
{

library(ape)
n <- 20
Lab <- paste("Node",1:(n-1),sep="")
tipLab <- paste("T",1:n,sep="")

# entire tree
Tree <- rtree(n,tip.label= tipLab)
new.Tree <- addNodeLab(treeO = Tree, nodeLab = Lab)
smallTree <- pruneTree(new.Tree)
pruneTree(Tree)
}
}
