% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pruneTree.R
\name{pruneTree}
\alias{pruneTree}
\title{prune tree at each node}
\usage{
pruneTree(tree, message = FALSE)
}
\arguments{
\item{wtree}{an object of phylo class}
}
\value{
a list of phylo object. If the tree ('wtree') provides node and tip labels, the labels are passed to its subtrees; otherwise, the labels are generated automatically with a prefix 'Node_' for internal nodes and 'Tip_' for leaves with the corresponding numbers in the edge component of phylo object. (wtree$edge)
}
\description{
\code{pruneTree} prune the tree at each node.
}
\examples{

library(ape)
n <- 20
tree<- rtree(n)
small <- pruneTree(tree, message = TRUE)
}
