% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmDesc.R
\name{rmDesc}
\alias{rmDesc}
\title{remove nodes which are the descendants of other nodes}
\usage{
rmDesc(node, tree)
}
\arguments{
\item{node}{a vector of node labels or node numbers}

\item{tree}{a list of phylo object.}
}
\value{
a vector of node numbers
}
\description{
\code{rmDesc} is to remove branches, which are the subbranch of the others.
}
\examples{
{
data(tinyTree)
library(ggtree)
# PLOT
ggtree(tinyTree,branch.length = 'none') +
  geom_text2(aes(label=node), hjust=-.3)

# find the shared nodes from the tree plot
aV <- c(5, 4, 18, 3, 2)
# final result
(rn <- rmDesc(node = aV, tree = tinyTree))
}

}
