% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmMember.R
\name{rmMember}
\alias{rmMember}
\title{remove nodes which are the descendants of other nodes}
\usage{
rmMember(node, tree)
}
\arguments{
\item{node}{a vector of node labels or node numbers}

\item{stree}{a list of phylo object.}
}
\value{
a vector of node numbers
}
\description{
\code{rmMember} is to remove branches, which are the subbranch of the others.
}
\examples{
{
library(ape)
n <- 20

# entire tree
set.seed(1)
Tree <- rtree(n)

# PLOT
ggtree(Tree,branch.length = 'none') +
  geom_text2(aes(label=node), hjust=-.3)

# find the shared nodes from the tree plot
aV <- c(36, 33, 37, 16)
# final result
rn <- rmMember(node = aV, tree = Tree)
}

}
