% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shareNode.R
\name{shareNode}
\alias{shareNode}
\title{Find the share node}
\usage{
shareNode(tree, node, label = FALSE)
}
\arguments{
\item{tree}{A  phylo object.}

\item{node}{A vector of node numbers or node labels.}

\item{label}{A logical value. If TRUE, the selected node label is output;
otherwise the node number is output}
}
\value{
The label of the shared node
}
\description{
\code{shareNode} is to find the node where the specified nodes
first meet.
}
\examples{

data(tinyTree)
library(ggtree)
# PLOT tree
ggtree(tinyTree,branch.length = 'none')+
geom_text2(aes(label = label))

## find the node shared by provided node labels
shareNode(node = c('t4','t9'), tree = tinyTree,
 label = TRUE)

shareNode(node = c('t10','Node_17'), tree = tinyTree,
 label = TRUE)

# -----------------------------------------------------
ggtree(tinyTree,branch.length = 'none')+
geom_text2(aes(label = node))

## find the node shared by provided node numbers
shareNode(node = c(2, 3), tree = tinyTree)
}
\author{
Ruizhu Huang
}
