% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shareNode.R
\name{shareNode}
\alias{shareNode}
\title{find the node shared by specified tips / nodes or both}
\usage{
shareNode(child, wtree, stree)
}
\arguments{
\item{child}{a vector of specified tips or/and nodes}

\item{wtree}{a tree (phylo class)}

\item{stree}{a list of subtrees which are cut at each node of the wtree}
}
\value{
the label of the shared node
}
\description{
\code{shareNode} could find the lowest level node which is shared
 by the specified descendants. (the root is the highest level on the tree)
}
