% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signalFind.R
\name{signalFind}
\alias{signalFind}
\title{Find the signal location (node/tip/both) in a tree}
\usage{
signalFind(TipDiff, wtree, stree)
}
\arguments{
\item{TipDiff}{the labels of significant differential tips}

\item{wtree}{the entire phylogenetic tree which is based to do the aggregation}

\item{stree}{a list of subtrees which are cut at the nodes of phylogenetic tree}
}
\value{
a character vector which shows the simplified result : signal nodes and tips
which can not be aggregated any more on the tree
}
\description{
Given the known significant differential tips,
this functions could find the highest nodes on the tree shared by these tips
and aggregate the tips by their shared nodes to make the result short
}
\examples{
\\notrun{
library(ape)
n <- 20
Lab <- paste("Node",1:(n-1),sep="")
tipLab <- paste("T",1:n,sep="")
# entire tree
Tree <- rtree(n,tip.label= tipLab)
Tree <- addNodeLab(treeO = Tree, nodeLab = Lab)

# Subtrees
sTree <- pruneTree(Tree)


# PLOT
ggtree(Tree,branch.length = "none") +
geom_text2(aes(subset=!isTip, label=label), hjust=-.3) +
geom_tiplab()

# Differential tips (expect to get "Node4" "Node8" "T1")
# find the shared nodes from the tree plot
DiffTip <- c("T7","T15","T1","T9","T19")

# final result
rn <- signalFind(DiffTip, Tree, sTree)
}
}
