% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signalNode.R
\name{signalNode}
\alias{signalNode}
\title{find the optimal nodes to short result.}
\usage{
signalNode(tree, node, label = FALSE)
}
\arguments{
\item{tree}{A tree (phylo object)}

\item{node}{A vector of node numbers or node labels}

\item{label}{A logical value. If TRUE, the selected node label is output;
otherwise the node number is output}
}
\value{
The label of the shared node
}
\description{
\code{signalNode} is to represent some nodes with their ancestor to make result
as short as possible. The ancestors share exactly the same leaves as the
original nodes.
}
\examples{

data(tinyTree)
library(ggtree)

# PLOT tree
ggtree(tinyTree,branch.length = 'none')+
geom_text2(aes(label = label))

## find the node shared by provided node labels
signalNode(node = c('t4','t9'), tree = tinyTree,
 label = TRUE)

signalNode(node = c('t10','Node_18', 't8'), tree = tinyTree,
 label = TRUE)

# -----------------------------------------------------
ggtree(tinyTree,branch.length = 'none')+
geom_text2(aes(label = node))

## find the node shared by provided node numbers
signalNode(node = c(2, 3), tree = tinyTree)
signalNode(node = c(2, 3, 16), tree = tinyTree)

}
\author{
Ruizhu Huang
}
