% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpr.R
\name{tpr}
\alias{tpr}
\title{Calculate true positive rate (TPR) on a tree structure}
\usage{
tpr(tree, truth, found, level = c("node", "leaf"), direction = FALSE)
}
\arguments{
\item{tree}{A phylo object}

\item{truth}{Nodes that have signals (eg. differentally abundant at different experimental conditions.). If the signals are in different directions (up and down), then provide the nodes as a list of two members (one up and one down). One could provide either node number or node label. \strong{Note:} When TPR at node level is required and only leaf nodes with signal are provided, then the internal nodes which are shared and only shared by the provied leaf nodes (signal in the same direction) will be found out and used with the leaf nodes in the TPR calculation; when TPR at leaf level is required and the given nodes have internal nodes, then the descendant leaf nodes will be found out and used in the TPR calculation.}

\item{found}{Nodes that have been found to have signal (eg. differentally abundant at different experimental conditions). If the signals are in different directions (up and down), then provide the nodes as a list of two members (one up and one down). One could provide either node number or node label. \strong{Note:} When TPR at node level is required, then the descendant nodes of the provied nodes (include themselves) will be found out and used in the TPR calculation; when TPR at leaf level is required, then the descendant leaf nodes will be found out and used in the TPR calculation.}

\item{level}{If "leaf", true positive rate is calculated at leaf level; if "node", it is calculated at node level.}

\item{direction}{TRUE or FALSE. Default is FALSE. If TRUE, the signal direction is taken into account; the argument \strong{truth} and \strong{found} should both be a list of two members and the order of directions should match.}
}
\value{
A true positive rate
}
\description{
\code{tpr} calculates the true positive rate (TPR) on a tree structure at leaf or node level.
}
\examples{

library(ggtree)
data("tinyTree")
ggtree(tinyTree) + geom_text2(aes(label = node))

# Truth: two branches have differential
# abundance under different conditions.
# branch with node 16 (direction: increase)
# branch with node 13 (direction: decrease)

# ------------ ignore direction -------------------
# Found: branches with node 17 and node 14
 # TPR at the tip level
tpr1 <- tpr(tree = tinyTree, truth = c(16, 13),
            found = c(17, 14), level = "leaf")
 # TPR at the node level
tpr2 <- tpr(tree = tinyTree, truth = c(16, 13),
            found = c(17, 14), level = "node")


# ------------ direction matters-------------------
# Found: branches with node 17 (up) and node 14 (down)

 # the order of direction for truth and found should be the same
(tpr3 <- tpr(tree = tinyTree, truth = list(16, 13),
            found = list(16, 13), level = "leaf",
            direction = TRUE))  # correct order

(tpr4 <- tpr(tree = tinyTree, truth = list(16, 13),
            found = list(13, 16), level = "node",
            direction = TRUE)) # wrong order

}
