% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transNode.R
\name{transNode}
\alias{transNode}
\title{Transfer between node number and node label}
\usage{
transNode(tree, input)
}
\arguments{
\item{tree}{A phylo object}

\item{input}{A tree node label or a tree node number}
}
\value{
a vector
}
\description{
\code{transNode} does the transformation between the number and the label of
a node on a tree
}
\examples{
library(ggtree)

data(tinyTree)

ggtree(tinyTree, branch.length = 'none') +
geom_text2(aes(label = label), hjust = -0.3) +
geom_text2(aes(label = node), vjust = -0.8,
hjust = -0.3, color = 'blue')

#check whether the node number and node label are matched
transNode(tinyTree, input = c(11, 2, 4))

transNode(tinyTree, input = c('Node_16', 'Node_11'))

}
\author{
Ruizhu Huang
}
