% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transNode.R
\name{transNode}
\alias{transNode}
\title{transfer between node number and node label}
\usage{
transNode(tree, input)
}
\arguments{
\item{tree}{a phylo object}

\item{input}{a tree node label or a tree node number}
}
\description{
\code{transNode} is to do the transfermation between the number and the label of a node on a tree
}
\examples{


library(ggtree)

data(tinyTree)

ggtree(tinyTree, branch.length = 'none') +
geom_text2(aes(label = label), hjust = -0.3) +
geom_text2(aes(label = node), vjust = -0.8,
hjust = -0.3, color = 'blue')

#check whether the node number and node label are matched
transNode(tinyTree, input = c(11, 2, 4))

transNode(tinyTree, input = c('Node_16', 'Node_11'))


}
