% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeAGG.R
\name{treeAGG}
\alias{treeAGG}
\title{Tree aggregation}
\usage{
treeAGG(tree, data, stree = NULL, P.lim = NULL, varSIG = NULL, varAGG)
}
\arguments{
\item{tree}{A phylo object;}

\item{data}{A data frame (include at least :
1. the label of tree nodes and tips as row names
2. a column of p value
3. a column of adjusted p value}

\item{stree}{A list of phylo object. The subtrees of \strong{tree}.}

\item{P.lim}{The threshold value (for the adjusted p values) to reject a null hypothesis. By default, NULL. If NULL, the algorithm only compares the value provided by \strong{varAGG} and doesn't decide whether to reject a null hypothesis.}

\item{varSIG}{The name of column for testing significance}

\item{varAGG}{The name of column used to do tree aggregation, eg. the name of the p value or adjusted p value column}
}
\value{
A character vector, containing the labels of tips and nodes which
have selected by the minimum P-value algorithm
}
\description{
\code{treeAGG} combines the p values with the tree structure and decide the which nodes to be aggregated to based on the min-p algorithm.
}
