% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minP_algorithm.R
\name{treeAGG}
\alias{treeAGG}
\title{Apply minimum P-value algorithm}
\usage{
treeAGG(wtree, ResTipNode, stree, P.lim, VarSig, VarAGG)
}
\arguments{
\item{wtree}{phylo object; the entire tree for all OTUs}

\item{ResTipNode}{data frame (include at least :
1. the label of tree nodes and tips as row names
2. adjusted-p value in one column)}

\item{stree}{a list of all the subtrees of a phylogenetic tree (wtree)}

\item{P.lim}{the threshold value for adjusted-P}

\item{VarSig}{column name for testing significance}

\item{VarAGG}{variable used to do tree aggregation}
}
\value{
a vector of character shows the labels of tips and nodes which
have filtered by the minimum P-value algorithm
}
\description{
This function applies minimum p-value algorithm to find the branch which
is differentially abundant.
}
\examples{
{

}

}
