% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/classAccessor.R
\docType{methods}
\name{assayNames,treeSummarizedExperiment-method}
\alias{assayNames,treeSummarizedExperiment-method}
\alias{assayNames<-,treeSummarizedExperiment,ANY-method}
\alias{assays,treeSummarizedExperiment-method}
\alias{assays<-,treeSummarizedExperiment,ANY-method}
\alias{rowData,treeSummarizedExperiment-method}
\alias{rowData<-,treeSummarizedExperiment-method}
\alias{colData,treeSummarizedExperiment-method}
\alias{colData<-,treeSummarizedExperiment,ANY-method}
\alias{metadata,treeSummarizedExperiment-method}
\alias{linkData}
\alias{linkData,treeSummarizedExperiment-method}
\alias{treeData}
\alias{treeData,treeSummarizedExperiment-method}
\alias{[,treeSummarizedExperiment,ANY,ANY,ANY-method}
\alias{treeSummarizedExperiment-accessor}
\title{Accessor functions for treeSummarizedExperiment}
\usage{
\S4method{assayNames}{treeSummarizedExperiment}(x, ...)

\S4method{assayNames}{treeSummarizedExperiment,ANY}(x, ...) <- value

\S4method{assays}{treeSummarizedExperiment}(x, use.nodeLab = FALSE,
  withDimnames = FALSE)

\S4method{assays}{treeSummarizedExperiment,ANY}(x) <- value

\S4method{rowData}{treeSummarizedExperiment}(x, ...)

\S4method{rowData}{treeSummarizedExperiment}(x, ...) <- value

\S4method{colData}{treeSummarizedExperiment}(x, ...)

\S4method{colData}{treeSummarizedExperiment,ANY}(x, ...) <- value

\S4method{metadata}{treeSummarizedExperiment}(x)

linkData(x)

\S4method{linkData}{treeSummarizedExperiment}(x)

treeData(x)

\S4method{treeData}{treeSummarizedExperiment}(x)

\S4method{[}{treeSummarizedExperiment,ANY,ANY,ANY}(x, i, j)
}
\arguments{
\item{x}{A treeSummarizedExperiment object}

\item{...}{For assay, ... contains \code{use.nodeLab}, which is forwarded to
assays. For rowData, arguments passed through ... are forwarded to mcols.
For cbind, rbind, ... contains leafSummarizedExperiment objects to be
combined. For other accessors, ignored.}

\item{value}{An object of a class specified in the S4 method signature. See
\code{\link[SummarizedExperiment]{SummarizedExperiment}} for more details.}

\item{use.nodeLab}{A logical(1), indicating whether the rownames of assay
elements should use node labels (the column \code{nodeLab} in
\code{linkData} if there is not duplicated values; otherwise the column
\code{nodeLab_allias} in \code{linkData} is used.)}

\item{withDimnames}{A logical(1), indicating whether dimnames should be
applied to extracted assay elements. Setting withDimnames=FALSE increases
the speed and memory efficiency with which assays are extracted.
withDimnames=TRUE in the getter assays<- allows efficient complex
assignments (e.g., updating names of assays, names(assays(x,
withDimnames=FALSE)) = ... is more efficient than names(assays(x)) = ...);
it does not influence actual assignment of dimnames to assays.}

\item{i, j}{The subscripts that can act to subset the rows and columns of
\code{x}, that is the matrix elements of assays.}
}
\value{
Elements from \code{treeSummarizedExperiment}.
}
\description{
Accessor functions to extract different elements from
\strong{treeSummarizedExperiment} object.
}
\seealso{
\code{\link{treeSummarizedExperiment}}
  \code{\link{treeSummarizedExperiment-accessor}}
  \code{\link{leafSummarizedExperiment}}
  \code{\link[SummarizedExperiment]{SummarizedExperiment-class}}
}
\author{
Ruizhu HUANG
}
